### 2022-11-30 #4985 @editorialbot
[PRE REVIEW]: Pyra: Automated EM27/SUN Greenhouse Gas Measurement Software
**Submitting author:** @patrickjaigner (<a href="http://orcid.org/0000-0002-1530-415X">Patrick Aigner</a>)
**Repository:** https://github.com/tum-esm/pyra
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** 4.0.6
