+++ README.md
@@ -2,7 +2,7 @@
 
 **Documentation on: [pyra.esm.ei.tum.de/docs](https://pyra.esm.ei.tum.de/docs)**
 
-**Contributor Guide on: [pyra.esm.ei.tum.de/docs/contributor-guide](https://pyra.esm.ei.tum.de/docs/contributor-guide)**
+**Contributor Guide on: [https://pyra.esm.ei.tum.de/docs/category/contributor-guide](https://pyra.esm.ei.tum.de/docs/category/contributor-guide)**
 
 [![GitHub Workflow Status](https://img.shields.io/github/workflow/status/tum-esm/pyra/test-python-codebase-on-main?label=python%20tests%20on%20main%20branch&color=86efac&style=for-the-badge)](https://github.com/tum-esm/pyra/actions/workflows/test-python-codebase-on-main.yml)
 [![GitHub Workflow Status](https://img.shields.io/github/workflow/status/tum-esm/pyra/test-typescript-codebase-on-main?label=typescript%20tests%20on%20main%20branch&color=86efac&style=for-the-badge)](https://github.com/tum-esm/pyra/actions/workflows/test-typescript-codebase-on-main.yml)<br/>
