+++ README.md
@@ -4,7 +4,8 @@
 
 **Contributor Guide on: [https://pyra.esm.ei.tum.de/docs/category/contributor-guide](https://pyra.esm.ei.tum.de/docs/category/contributor-guide)**
 
-[![GitHub Workflow Status](https://img.shields.io/github/workflow/status/tum-esm/pyra/test-python-codebase-on-main?label=python%20tests%20on%20main%20branch&color=86efac&style=for-the-badge)](https://github.com/tum-esm/pyra/actions/workflows/test-python-codebase-on-main.yml)
-[![GitHub Workflow Status](https://img.shields.io/github/workflow/status/tum-esm/pyra/test-typescript-codebase-on-main?label=typescript%20tests%20on%20main%20branch&color=86efac&style=for-the-badge)](https://github.com/tum-esm/pyra/actions/workflows/test-typescript-codebase-on-main.yml)<br/>
+![GitHub Workflow Status](https://img.shields.io/github/actions/workflow/status/tum-esm/pyra/test-python-codebase-on-main.yml?branch=main&color=86efac&label=python%20tests%20on%20main%20branch&style=for-the-badge)
+![GitHub Workflow Status](https://img.shields.io/github/actions/workflow/status/tum-esm/pyra/test-typescript-codebase-on-main.yml?branch=main&color=86efac&label=typescript%20tests%20on%20main%20branch&style=for-the-badge)
+<br/>
 [![GitHub release (latest by date)](https://img.shields.io/github/v/release/tum-esm/pyra?display_name=tag&label=latest%20release&color=fcd34d&style=for-the-badge)](https://github.com/tum-esm/pyra/releases)
 [![GitHub](https://img.shields.io/github/license/tum-esm/pyra?color=fcd34d&style=for-the-badge)](https://github.com/tum-esm/pyra/blob/main/LICENSE.md)
