### 2023-01-04 006aff9
Reimplement ring list
