### 2023-01-04 689e999
Debug and expand ring list
