### 2023-01-04 68d0af0
Expand ring list tests
