### 2023-01-04 aefc9fa
#147 - make helios queue use edge fractions directly
