### 2023-01-04 6cd1d21
#147 - remove unused config param from core
