### 2023-01-04 208c7dd
#147 - remove unused config param from ui
