### 2023-01-04 8b9f409
#147 - remove unused config param from docs
enforce min max on edge_detection_threshold
