### 2023-01-23 #4985 @kthyng
@patrickjaigner thank you for your patience! We are short on editors right now, adding to delays.

In the meantime could you provide a list (5+) of suggested reviewers? They can be volunteers from [this list](https://bit.ly/joss-reviewers) or people you know of, as long as they are not conflicted. Please don't "@" them to avoid pinging them with github notifications preemptively.
