+++ packages/core/utils/helios_image_processing.py
@@ -1,6 +1,7 @@
 from datetime import datetime
+import math
 import os
-from typing import Literal
+from typing import Optional
 import cv2 as cv
 import numpy as np
 
@@ -9,12 +10,15 @@ PROJECT_DIR = dir(dir(dir(dir(os.path.abspath(__file__)))))
 IMG_DIR = os.path.join(PROJECT_DIR, "logs", "helios")
 
 
-class ImageProcessing:
+class HeliosImageProcessing:
 
     # circle code adapted from https://stackoverflow.com/a/39074620/8255842
     @staticmethod
     def _get_circle_mask(
-        img_shape: tuple[int, int], radius: int, center_x: int, center_y: int
+        img_shape: tuple[int, int],
+        radius: int,
+        center_x: int,
+        center_y: int,
     ) -> cv.Mat:
         """
         input: image width/height, circle radius/center_x/center_y
@@ -35,7 +39,10 @@ class ImageProcessing:
         return (np.abs(np.hypot(center_x - x, center_y - y)) < radius).astype(np.uint8)
 
     @staticmethod
-    def _moving_average(xs: list[float], n: int = 3) -> list[float]:
+    def _moving_average(
+        xs: list[float],
+        n: int = 3,
+    ) -> list[float]:
         ret = np.cumsum(xs)
         ret[n:] = ret[n:] - ret[:-n]
         return list(ret[n - 1 :] / n)
@@ -69,7 +76,7 @@ class ImageProcessing:
         return binary_mask
 
     @staticmethod
-    def _get_circle_location(binary_mask: cv.Mat) -> tuple[int, int, int]:
+    def _get_circle_location(grayscale_image: cv.Mat) -> Optional[tuple[int, int, int]]:
         """
         input: binary mask (2D array) like
         [[0 0 0 0 0 0 0 0 0 0 0 0 0 0 0 0]
@@ -82,32 +89,38 @@ class ImageProcessing:
         [0 0 0 0 0 1 1 1 1 1 1 1 0 0 0 0]
         [0 0 0 0 0 0 0 1 1 1 0 0 0 0 0 0]]
 
-        output: center_x, center_y, r (all floats)
+        output: center_x, center_y, r
         """
+        image_height, image_width = grayscale_image.shape[0], grayscale_image.shape[1]
 
-        # number of bright pixels in each image row
-        row_sums = np.sum(binary_mask, axis=1)
+        circles = cv.HoughCircles(
+            grayscale_image,
+            cv.HOUGH_GRADIENT,
+            dp=20,
+            minDist=250,
+            minRadius=round((image_height * 0.85) * 0.5),
+            maxRadius=round((image_height * 1.15) * 0.5),
+        )
 
-        # WxH mask: 0 in black pixels, column-index in bright pixels
+        if circles is None:
+            return None
 
-        row_indices, col_indices = np.indices(binary_mask.shape)
-        row_indices *= binary_mask
-        col_indices *= binary_mask
-
-        radius = np.max(ImageProcessing._moving_average(row_sums, n=3)) / 2
-        center_y = np.mean(row_indices, where=(row_indices != 0))
-        center_x = np.mean(col_indices, where=(col_indices != 0))
-
-        return round(center_x), round(center_y), round(radius)
+        # get the circle that is closest to the image center
+        x, y, r = min(
+            np.round(circles[0, :]).astype("int"),
+            key=lambda c: math.pow(c[0] - (image_width * 0.5), 2)  # type: ignore
+            + pow(c[1] - (image_height * 0.5), 2),
+        )
+        return round(x), round(y), round(r)
 
     @staticmethod
-    def _add_markings_to_image(
+    def add_markings_to_image(
[…]
