+++ packages/core/utils/__init__.py
@@ -3,5 +3,5 @@ from .ring_list import RingList
 from .astronomy import Astronomy
 from .exception_email_client import ExceptionEmailClient
 from .update_dict_recursively import update_dict_recursively
-from .image_processing import ImageProcessing
+from .helios_image_processing import HeliosImageProcessing
 from .with_filelock import with_filelock
