+++ pyproject.toml
@@ -25,9 +25,6 @@ tqdm = "^4.64.0"
 colorama = "^0.4.5"
 deepdiff = "^6.2.1"
 
-[tool.poetry.group.dev]
-optional = true
-
 # libraries only used in development
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.1.2"
@@ -49,6 +46,7 @@ line-length = 95
 strict = true
 implicit_reexport = true
 no_warn_unused_ignores = true
+plugins = ["pydantic.mypy"]
 
 [[tool.mypy.overrides]]
 module = [
