### 2023-02-14 5e94138
Refine static type checks script
