+++ .github/workflows/test-python-codebase-on-main.yml
@@ -31,19 +31,6 @@ jobs:
             # check-out repo and set-up python
             - name: Check out repository
               uses: actions/checkout@v3
-            - name: Set up Python 3.10.6
-              id: setup-python
-              uses: actions/setup-python@v3
-              with:
-                  python-version: 3.10.6
-
-            # install & configure poetry
-            - name: Install Poetry
-              uses: snok/install-poetry@v1
-              with:
-                  virtualenvs-create: true
-                  virtualenvs-in-project: true
-                  installer-parallel: true
 
             # load cached venv if cache exists
             - name: Load cached venv
@@ -51,18 +38,32 @@ jobs:
               uses: actions/cache@v3
               with:
                   path: .venv
-                  key: venv-${{ runner.os }}-${{ steps.setup-python.outputs.python-version }}-${{ hashFiles('poetry.lock') }}
+                  key: venv-${{ runner.os }}-3.10.6-${{ hashFiles('poetry.lock') }}
+
+            # set up venv if not cached
+            - name: Set up Python 3.10.6
+              if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
+              uses: actions/setup-python@v3
+              with:
+                  python-version: 3.10.6
+            - name: Install Poetry
+              if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
+              uses: snok/install-poetry@v1
+              with:
+                  virtualenvs-create: true
+                  virtualenvs-in-project: true
+                  installer-parallel: true
 
             # install dependencies if cache does not exist
             - name: Install dependencies
               if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
-              run: poetry install --no-interaction --with dev,helios
+              run: poetry install
 
             # run test suite
             - name: Run mypy static type analysis
               run: |
                   source .venv/bin/activate
-                  bash scripts/run_type_analysis.sh
+                  bash scripts/check_static_types.sh
             - name: Run pytest tests
               run: |
                   source .venv/bin/activate
