+++ packages/ui/src/custom-types.ts
@@ -39,7 +39,7 @@ export namespace customTypes {
         | 'upload';
     export type config = {
         general: {
-            version: '4.0.5';
+            version: '4.0.7';
             seconds_per_core_interval: number;
             test_mode: boolean;
             station_id: string;
@@ -108,7 +108,7 @@ export namespace customTypes {
     // I have not found a more elegant way yet to generate a partialConfig type
     export type partialConfig = {
         general?: {
-            version?: '4.0.5';
+            version?: '4.0.7';
             seconds_per_core_interval?: number;
             test_mode?: boolean;
             station_id?: string;
