+++ .github/workflows/build-on-release.yml
@@ -13,14 +13,27 @@ jobs:
         runs-on: windows-latest
         steps:
             - uses: actions/checkout@v2
-            - name: setup node
-              uses: actions/setup-node@v1
+
+            - name: Set up NodeJS with Yarn
+              uses: actions/setup-node@v3
               with:
                   node-version: 16
+                  cache: 'yarn'
+                  cache-dependency-path: yarn.lock
+
             - name: install Rust stable
               uses: actions-rs/toolchain@v1
               with:
                   toolchain: stable
+
+            # load cached tauri stuff
+            - name: Load cached tauri stuff
+              id: cached-poetry-dependencies
+              uses: actions/cache@v3
+              with:
+                  path: src-tauri/target
+                  key: tauri-stuff-${{ runner.os }}
+
             - name: install app dependencies and build it
               run: yarn && yarn build
             - uses: tauri-apps/tauri-action@v0
