+++ .github/workflows/test-python-codebase-on-main.yml
@@ -33,30 +33,31 @@ jobs:
               uses: actions/checkout@v3
 
             # load cached venv if cache exists
-            - name: Load cached venv
-              id: cached-poetry-dependencies
-              uses: actions/cache@v3
-              with:
-                  path: .venv
-                  key: venv-${{ runner.os }}-3.10.6-${{ hashFiles('poetry.lock') }}
+            # - name: Load cached venv
+            #   id: cached-poetry-dependencies
+            #   uses: actions/cache@v3
+            #   with:
+            #       path: .venv
+            #       key: venv-${{ runner.os }}-3.10.6-${{ hashFiles('poetry.lock') }}
 
             # set up venv if not cached
             - name: Set up Python 3.10.6
-              if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
+              #if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
               uses: actions/setup-python@v3
               with:
                   python-version: 3.10.6
             - name: Install Poetry
-              if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
+              #if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
               uses: snok/install-poetry@v1
               with:
+                  version: 1.3.2
                   virtualenvs-create: true
                   virtualenvs-in-project: true
                   installer-parallel: true
 
             # install dependencies if cache does not exist
             - name: Install dependencies
-              if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
+              #if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
               run: poetry install
 
             # run test suite
