### 2023-02-14 dd81922
Debug Helios thread
