+++ pyproject.toml
@@ -8,33 +8,33 @@ authors = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.10"
-Cerberus = "^1.3.4"
-numpy = "^1.24.2"
-filelock = "^3.10.7"
-pywin32 = { version = "^303", markers = "sys_platform == 'win32'"}
-click = "^8.1.3"
-python-snap7 = "^1.3"
-opencv-python = "^4.7.0.72"
-astropy = "^5.2.2"
-jdcal = "^1.4.1"
+python = "3.10.*"
+Cerberus = "1.3.4"
+numpy = "1.24.2"
+filelock = "3.10.7"
+pywin32 = { version = "303", markers = "sys_platform == 'win32'"}
+click = "8.1.3"
+python-snap7 = "1.3"
+opencv-python = "4.7.0.72"
+astropy = "5.2.2"
+jdcal = "1.4.1"
 psutil = "5.9.4"
-fabric = "^2.7.1"
-pydantic = "^1.10.7"
-tqdm = "^4.65.0"
-colorama = "^0.4.6"
-deepdiff = "^6.3.0"
-tum-esm-utils = "^1.3.0"
+fabric = "2.7.1"
+pydantic = "1.10.7"
+tqdm = "4.65.0"
+colorama = "0.4.6"
+deepdiff = "6.3.0"
+tum-esm-utils = "1.3.0"
 
 [tool.poetry.group.dev.dependencies]
-pytest = "^7.2.2"
-black = "^22.12.0"
-types-paramiko = "^2.12.0"
-types-invoke = "^1.7.3"
-pytest-cov = "^3.0.0"
-types-psutil = "^5.9.5.11"
-lazydocs = "^0.4.8"
-mypy = "^1.1.1"
+pytest = "7.2.2"
+black = "22.12.0"
+mypy = "1.1.1"
+types-paramiko = "2.12.0.3"
+types-invoke = "1.7.3.17"
+pytest-cov = "3.0.0"
+types-psutil = "5.9.5.11"
+lazydocs = "0.4.8"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
