### 2023-02-17 9cb740a
Get rid of commit sha in emails
