+++ packages/core/utils/exception_email_client.py
@@ -3,9 +3,7 @@ import os
 import smtplib, ssl
 from email.mime.text import MIMEText
 from email.mime.multipart import MIMEMultipart
-import subprocess
 import traceback
-from typing import Optional
 from packages.core import types
 
 _dir = os.path.dirname
@@ -19,30 +17,6 @@ def _get_pyra_version() -> str:
         pyra_version: str = json.load(f)["version"]
     assert pyra_version.startswith("4.")
     return pyra_version
-
-
-# TODO: get rid of commit sha
-
-
-def _get_commit_sha() -> Optional[str]:
-    """Get the current commit sha of the PYRA codebase
-    Returns None if git is not installed or directory is
-    a git repository."""
-
-    commit_sha_process = subprocess.run(
-        ["git", "rev-parse", "--verify", "HEAD", "--short"],
-        stdout=subprocess.PIPE,
-        stderr=subprocess.STDOUT,
-    )
-    if commit_sha_process.returncode == 0:
-        stdout = commit_sha_process.stdout.decode()
-        commit_sha = stdout.replace("\n", "").replace(" ", "")
-        assert len(commit_sha) > 0
-        if "fatal: not a git repository" in stdout:
-            return None
-        return commit_sha
-    else:
-        return None
 
 
 def _get_current_log_lines() -> list[str]:
@@ -145,7 +119,6 @@ class ExceptionEmailClient:
             return
 
         pyra_version = _get_pyra_version()
-        commit_sha = _get_commit_sha()
         current_log_lines = _get_current_log_lines()
 
         tb = "\n".join(traceback.format_exception(exception))
@@ -154,9 +127,7 @@ class ExceptionEmailClient:
         text = (
             f"{type(exception).__name__} has occured. Details:\n"
             + f"{tb}\nLast 2 iteration's log lines:{logs}\n"
-            + f"This email has been generated by Pyra {pyra_version} "
-            + (f"(commit {commit_sha}) " if commit_sha is not None else "")
-            + "automatically."
+            + f"This email has been generated by Pyra {pyra_version}."
         )
 
         pre_tag = '<pre style="background-color: #f1f5f9; color: #334155; padding: 8px 8px 12px 12px; border-radius: 3px; max-width: full; overflow-x: scroll;">'
@@ -168,7 +139,7 @@ class ExceptionEmailClient:
             + f"    {pre_tag}<code>{tb}</code></pre>"
             + "    <p><strong>Last 2 iteration's log lines:</strong></p>"
             + f"    {pre_tag}<code>{logs}</code></pre>"
-            + f"    <p><em>This email has been generated by Pyra {pyra_version} (commit {commit_sha}) automatically.</em></p>"
+            + f"    <p><em>This email has been generated by Pyra {pyra_version}.</em></p>"
             + "  </body>"
             + "</html>"
         )
