### 2023-02-28 #5131 @patrickjaigner
Thank you for your feedback @arthur-e .

We have already updated your points on the software paper. (Line 8-9, Line 28, Line 40, Line 41, Line 72, Line 99).

To give a better understanding of the overall system setup we created an overview and added it to the website (https://pyra.esm.ei.tum.de/docs/user-guide/functionality). The usual setup involves a Windows-based machine, an Ethernet + USB connection, and the EM27/SUN. In addition, a weather-protected enclosure needs to be present to be able to leave the instrument deployed at the measurement location. This weather-protected enclosure can be the TUM enclosure or any other custom-made enclosure by the community.

Measurements with an EM27/SUN require to run CamTracker, a software that tracks the sun position and operates the mirrors accordingly, and OPUS, a software that sends commands to the EM27/SUN and collects the output data into interferogram files. Both CamTracker and OPUS are Windows based and are used to operate the instrument.

Pyra allows to start and stop measurements with a single command through interaction with CamTracker and OPUS and monitors that both softwares act in the way they are supposed to. A common error for CamTracker is, for example, the loss of its tracking target (sun), which leads to no sunlight available and measurements failing for the rest of the day. In the named example, Pyra would detect that the tracker is not fixed on the sun anymore and restarts the CamTracker tracking.

Without Pyra operators would need to remotely connect to the windows machine at the start of the day (if a remote connection is available for the measurement location), initialize CamTracker, load a macro into OPUS, and start measurements. During the day, the operator would need to check the system to ensure that measurements are running and that no measurement-blocking issues are present.

Lines 53-61: Sun Elevation and Time can be set as the operator sees fit (domain knowledge). In our experience, a sun angle under 10° produces measurement files that are unusable in the post-processing. Sun Elevations above 10° are no problem. If multiple triggers are selected i.e. Time and Sun Elevation an AND connection is considered and both conditions need to be fulfilled for a measurement to start. We offer both Time and Sun Elevation for the operator to customize the measurement period if no auxiliary sensor information about the sun or cloud coverage is present. (https://pyra.esm.ei.tum.de/docs/user-guide/measurements/#measurement-modes-and-triggers)

Addition info: I will be on my annual vacation from 03.03 until 24.03. During this time my co-authors will be available to answer questions and react to blocking issues.
