### 2023-02-28 #5131 @arthur-e
@patrickjaigner Thanks for the clarification.

Regarding the "Documentation" checklist above...

- Can you point me to the API documentation? [I see very nice, detailed guides for developers,](https://pyra.esm.ei.tum.de/docs/category/developer-guide/) but JOSS expects that a complete API reference also be made available.

@AoifeHughes After the API documentation is found, I've taken my review as far as I can. I don't have access to Windows (or OPUS, or CamTracker) so I can't verify any of the items under "Functionality."
