### 2023-03-01 #5131 @patrickjaigner
@arthur-e We didn't produce an explicit API documentation yet as we hoped that the source code and the CLI documentation would already act as comprehensive API documentation. We can create one if needed.
