### 2023-03-14 #5131 @arthur-e
Hi @AoifeHughes I tagged you earlier--After the API documentation is found, I've taken my review as far as I can. I don't have access to Windows (or OPUS, or CamTracker) so I can't verify any of the items under "Functionality."
