+++ packages/__init__.py
@@ -1,1 +1,1 @@
-from . import core
+from . import core, cli
