+++ packages/cli/commands/state.py
@@ -3,20 +3,20 @@ import click
 import os
 from packages.core import utils, interfaces
 
-dir = os.path.dirname
-PROJECT_DIR = dir(dir(dir(dir(os.path.abspath(__file__)))))
-STATE_FILE_PATH = os.path.join(PROJECT_DIR, "runtime-data", "state.json")
-STATE_LOCK_PATH = os.path.join(PROJECT_DIR, "config", ".state.lock")
+_dir = os.path.dirname
+_PROJECT_DIR = _dir(_dir(_dir(_dir(os.path.abspath(__file__)))))
+_STATE_FILE_PATH = os.path.join(_PROJECT_DIR, "runtime-data", "state.json")
+_STATE_LOCK_PATH = os.path.join(_PROJECT_DIR, "config", ".state.lock")
 
 
 @click.command(help="Read the current state.json file.")
 @click.option("--indent", is_flag=True, help="Print the JSON in an indented manner")
-@utils.with_filelock(STATE_LOCK_PATH)
+@utils.with_filelock(_STATE_LOCK_PATH)
 def _get_state(indent: bool) -> None:
-    if not os.path.isfile(STATE_FILE_PATH):
+    if not os.path.isfile(_STATE_FILE_PATH):
         interfaces.StateInterface.initialize()
 
-    with open(STATE_FILE_PATH, "r") as f:
+    with open(_STATE_FILE_PATH, "r") as f:
         try:
             state_content = json.load(f)
         except:
