+++ packages/cli/commands/core.py
@@ -6,18 +6,18 @@ import psutil
 from packages.core import utils, interfaces, modules
 
 dir = os.path.dirname
-PROJECT_DIR = dir(dir(dir(dir(os.path.abspath(__file__)))))
-INTERPRETER_PATH = (
-    "python" if os.name != "posix" else os.path.join(PROJECT_DIR, ".venv", "bin", "python")
+_PROJECT_DIR = dir(dir(dir(dir(os.path.abspath(__file__)))))
+_INTERPRETER_PATH = (
+    "python" if os.name != "posix" else os.path.join(_PROJECT_DIR, ".venv", "bin", "python")
 )
-CORE_SCRIPT_PATH = os.path.join(PROJECT_DIR, "run_pyra_core.py")
+_CORE_SCRIPT_PATH = os.path.join(_PROJECT_DIR, "run_pyra_core.py")
 
 
-def print_green(text: str) -> None:
+def _print_green(text: str) -> None:
     click.echo(click.style(text, fg="green"))
 
 
-def print_red(text: str) -> None:
+def _print_red(text: str) -> None:
     click.echo(click.style(text, fg="red"))
 
 
@@ -25,7 +25,7 @@ def process_is_running() -> Optional[int]:
     for p in psutil.process_iter():
         try:
             arguments = p.cmdline()
-            if (len(arguments) == 2) and (arguments[1] == CORE_SCRIPT_PATH):
+            if (len(arguments) == 2) and (arguments[1] == _CORE_SCRIPT_PATH):
                 return p.pid
         except (psutil.AccessDenied, psutil.ZombieProcess, psutil.NoSuchProcess):
             pass
@@ -38,7 +38,7 @@ def terminate_processes() -> list[int]:
         try:
             arguments = p.cmdline()
             if len(arguments) > 0:
-                if arguments[-1] == CORE_SCRIPT_PATH:
+                if arguments[-1] == _CORE_SCRIPT_PATH:
                     termination_pids.append(p.pid)
                     p.terminate()
         except (psutil.AccessDenied, psutil.ZombieProcess, psutil.NoSuchProcess):
@@ -52,15 +52,15 @@ def terminate_processes() -> list[int]:
 def _start_pyra_core() -> None:
     existing_pid = process_is_running()
     if existing_pid is not None:
-        print_red(f"Background process already exists with PID {existing_pid}")
+        _print_red(f"Background process already exists with PID {existing_pid}")
     else:
         p = subprocess.Popen(
-            [INTERPRETER_PATH, CORE_SCRIPT_PATH],
+            [_INTERPRETER_PATH, _CORE_SCRIPT_PATH],
             stdout=subprocess.PIPE,
             stderr=subprocess.PIPE,
         )
         utils.Logger.log_activity_event("start-core")
-        print_green(f"Started background process with PID {p.pid}")
+        _print_green(f"Started background process with PID {p.pid}")
 
 
 @click.command(
@@ -69,9 +69,9 @@ def _start_pyra_core() -> None:
 def _stop_pyra_core() -> None:
     termination_pids = terminate_processes()
     if len(termination_pids) == 0:
-        print_red("No active process to be terminated")
+        _print_red("No active process to be terminated")
     else:
-        print_green(
+        _print_green(
             f"Terminated {len(termination_pids)} pyra-core background "
             + f"processe(s) with PID(s) {termination_pids}"
         )
@@ -79,7 +79,7 @@ def _stop_pyra_core() -> None:
 
         config = interfaces.ConfigInterface.read()
         if config["general"]["test_mode"]:
-            print_green("Skipping TUM_PLC, CamTracker, and OPUS in test mode")
[…]
