### 2023-03-21 c70043e
Debug type check script
