### 2023-03-27 tum-esm/pyra#158 @nmstreethran
JOSS review
Hi, I am one of the reviewers assigned to your [submission to JOSS](https://github.com/openjournals/joss-reviews/issues/5131). My apologies for the delay in reviewing Pyra. Please find some comments regarding your submission below.

### Functionality / Documentation

- In the Setup guide (https://pyra.esm.ei.tum.de/docs/user-guide/setup):
  - There is a typo in the title: "**Hot** to set it up".
  - Can you add some information about how one can acquire and install OPUS/CamTracker? Are these software packages freely accessible? Provide links to their websites, if available.
- The [pyra-setup-tool](https://github.com/tum-esm/pyra-setup-tool) repo does not contain an open-source licence file.
- I suggest copying and pasting "What is Pyra?" from your documentation (https://pyra.esm.ei.tum.de/docs/) to this repo's README. The README currently only have links but having a brief overview of the software here would be good.

### Paper

- Statement of need
  - define GHG acronym in the second sentence
  - the acronym FTIR is used later in the text but isn't defined; this should probably be mentioned here?: "Ground-based remote sensing instruments that analyze direct sunlight using, for example, EM27/SUN Fourier transform spectrometers..."
- Summary
  - define the acronyms CLI, UI
- Bibliography: there are some formatting issues with the reference list in the PDF:
  - Most titles / book titles are not capitalised properly, e.g. see Che et al. (2022) where "beijing" is not capitalised; you can enclose the titles in double curly braces in the BibTeX file to fix this
  - Luther et al. (2022) - missing DOI: `10.5194/acp-22-5859-2022`
  - Kille et al. (2019) - `https://doi.org/` appears twice
  - Che et al. (2022) - subscript rendered incorrectly in the title (CO2)
  - Alberti et al. (2021) - the final author appears as "EM27/SUN-partners team, the"
  - Some author names have incorrect accents, for example Butz et al. (2017) (L"ubcke)
  - Journal titles are not consistent; most use full names but some use just the acronyms, for example Butz et al. (2017)

Many thanks!
