+++ packages/docs/docs/api-reference/ui/modules.md
@@ -1,0 +1,13 @@
+---
+id: "modules"
+title: "pyra-ui"
+sidebar_label: "Table of Contents"
+sidebar_position: 0.5
+hide_table_of_contents: true
+custom_edit_url: null
+---
+
+## Modules
+
+- [custom-types](modules/custom_types.md)
+- [utils](modules/utils.md)
