+++ packages/docs/docs/api-reference/ui/namespaces/custom_types.customTypes.md
@@ -1,0 +1,417 @@
+---
+id: "custom_types.customTypes"
+title: "Namespace: customTypes"
+sidebar_label: "customTypes"
+custom_edit_url: null
+---
+
+[custom-types](../modules/custom_types.md).customTypes
+
+## Type Aliases
+
+### OSState
+
+Ƭ **OSState**: `Object`
+
+#### Type declaration
+
+| Name | Type |
+| :------ | :------ |
+| `cpu_usage` | `number` \| ``null`` |
+| `filled_disk_space_fraction` | `number` \| ``null`` |
+| `last_boot_time` | `string` \| ``null`` |
+| `memory_usage` | `number` \| ``null`` |
+
+#### Defined in
+
+[custom-types.ts:255](https://github.com/tum-esm/pyra/blob/8f435c2/packages/ui/src/custom-types.ts#L255)
+
+___
+
+### activityHistory
+
+Ƭ **activityHistory**: { `event`: ``"start-core"`` \| ``"stop-core"`` \| ``"start-measurements"`` \| ``"stop-measurements"`` \| ``"error-occured"`` \| ``"errors-resolved"`` ; `localTime`: `string`  }[]
+
+#### Defined in
+
+[custom-types.ts:274](https://github.com/tum-esm/pyra/blob/8f435c2/packages/ui/src/custom-types.ts#L274)
+
+___
+
+### config
+
+Ƭ **config**: `Object`
+
+#### Type declaration
+
+| Name | Type |
+| :------ | :------ |
+| `camtracker` | { `config_path`: `string` ; `executable_path`: `string` ; `learn_az_elev_path`: `string` ; `motor_offset_threshold`: `number` ; `sun_intensity_path`: `string`  } |
+| `camtracker.config_path` | `string` |
+| `camtracker.executable_path` | `string` |
+| `camtracker.learn_az_elev_path` | `string` |
+| `camtracker.motor_offset_threshold` | `number` |
+| `camtracker.sun_intensity_path` | `string` |
+| `error_email` | { `notify_recipients`: `boolean` ; `recipients`: `string` ; `sender_address`: `string` ; `sender_password`: `string`  } |
+| `error_email.notify_recipients` | `boolean` |
+| `error_email.recipients` | `string` |
+| `error_email.sender_address` | `string` |
+| `error_email.sender_password` | `string` |
+| `general` | { `min_sun_elevation`: `number` ; `seconds_per_core_interval`: `number` ; `station_id`: `string` ; `test_mode`: `boolean` ; `version`: ``"4.0.7"``  } |
+| `general.min_sun_elevation` | `number` |
+| `general.seconds_per_core_interval` | `number` |
+| `general.station_id` | `string` |
+| `general.test_mode` | `boolean` |
+| `general.version` | ``"4.0.7"`` |
+| `helios` | ``null`` \| { `camera_id`: `number` ; `edge_detection_threshold`: `number` ; `evaluation_size`: `number` ; `save_images`: `boolean` ; `seconds_per_interval`: `number`  } |
+| `measurement_decision` | { `cli_decision_result`: `boolean` ; `manual_decision_result`: `boolean` ; `mode`: ``"automatic"`` \| ``"manual"`` \| ``"cli"``  } |
+| `measurement_decision.cli_decision_result` | `boolean` |
+| `measurement_decision.manual_decision_result` | `boolean` |
+| `measurement_decision.mode` | ``"automatic"`` \| ``"manual"`` \| ``"cli"`` |
[…]
