+++ packages/docs/docs/api-reference/ui/modules/assets.md
@@ -1,0 +1,29 @@
+---
+id: "assets"
+title: "Module: assets"
+sidebar_label: "assets"
+sidebar_position: 0
+custom_edit_url: null
+---
+
+## Variables
+
+### ICONS
+
+• `Const` **ICONS**: `Object`
+
+#### Type declaration
+
+| Name | Type |
+| :------ | :------ |
+| `alert` | `Element` |
+| `check` | `Element` |
+| `close` | `Element` |
+| `forbid` | `Element` |
+| `info` | `Element` |
+| `refresh` | `Element` |
+| `spinner` | `Element` |
+
+#### Defined in
+
+[assets/icons.tsx:19](https://github.com/tum-esm/pyra/blob/8f435c2/packages/ui/src/assets/icons.tsx#L19)
