+++ packages/docs/docs/api-reference/ui/modules/custom_types.md
@@ -1,0 +1,11 @@
+---
+id: "custom_types"
+title: "Module: custom-types"
+sidebar_label: "custom-types"
+sidebar_position: 0
+custom_edit_url: null
+---
+
+## Namespaces
+
+- [customTypes](../namespaces/custom_types.customTypes.md)
