+++ packages/docs/docs/api-reference/ui/modules/tabs.md
@@ -1,0 +1,77 @@
+---
+id: "tabs"
+title: "Module: tabs"
+sidebar_label: "tabs"
+sidebar_position: 0
+custom_edit_url: null
+---
+
+## Functions
+
+### AutomationTab
+
+▸ **AutomationTab**(): `Element`
+
+#### Returns
+
+`Element`
+
+#### Defined in
+
+[tabs/automation-tab.tsx:4](https://github.com/tum-esm/pyra/blob/8f435c2/packages/ui/src/tabs/automation-tab.tsx#L4)
+
+___
+
+### ConfigurationTab
+
+▸ **ConfigurationTab**(): `Element`
+
+#### Returns
+
+`Element`
+
+#### Defined in
+
+[tabs/configuration-tab.tsx:19](https://github.com/tum-esm/pyra/blob/8f435c2/packages/ui/src/tabs/configuration-tab.tsx#L19)
+
+___
+
+### ControlTab
+
+▸ **ControlTab**(): `Element`
+
+#### Returns
+
+`Element`
+
+#### Defined in
+
+[tabs/control-tab.tsx:89](https://github.com/tum-esm/pyra/blob/8f435c2/packages/ui/src/tabs/control-tab.tsx#L89)
+
+___
+
+### LogTab
+
+▸ **LogTab**(): `Element`
+
+#### Returns
+
+`Element`
+
+#### Defined in
+
+[tabs/log-tab.tsx:8](https://github.com/tum-esm/pyra/blob/8f435c2/packages/ui/src/tabs/log-tab.tsx#L8)
+
+___
+
+### OverviewTab
+
+▸ **OverviewTab**(): `Element`
+
+#### Returns
+
+`Element`
+
+#### Defined in
+
+[tabs/overview-tab.tsx:18](https://github.com/tum-esm/pyra/blob/8f435c2/packages/ui/src/tabs/overview-tab.tsx#L18)
