+++ packages/docs/docs/api-reference/cli/_category_.yml
@@ -1,0 +1,2 @@
+label: 'cli'
+position: 2
