+++ packages/docs/docs/api-reference/ui/_category_.yml
@@ -1,0 +1,2 @@
+label: "ui"
+position: 3
