+++ packages/docs/docs/api-reference/core/_category_.yml
@@ -1,0 +1,2 @@
+label: 'core'
+position: 1
