+++ packages/docs/docs/api-reference/ui/modules/components.md
@@ -1,110 +1,0 @@
----
-id: "components"
-title: "Module: components"
-sidebar_label: "components"
-sidebar_position: 0
-custom_edit_url: null
----
-
-## Variables
-
-### automationComponents
-
-• **automationComponents**: `Object`
-
-#### Type declaration
-
-| Name | Type |
-| :------ | :------ |
-| `MeasurementDecisionStatus` | () => `Element` |
-| `PyraCoreStatus` | () => `Element` |
-
-#### Defined in
-
-[components/automation/index.ts:4](https://github.com/tum-esm/pyra/blob/8f435c2/packages/ui/src/components/automation/index.ts#L4)
-
-___
-
-### configurationComponents
-
-• **configurationComponents**: `Object`
-
-#### Type declaration
-
-| Name | Type |
-| :------ | :------ |
-| `ConfigElementText` | (`props`: { `disabled?`: `boolean` ; `numeric?`: `boolean` ; `oldValue`: `string` \| `number` ; `postfix?`: `string` ; `showFileSelector?`: `boolean` ; `title`: `string` ; `value`: `string` \| `number` ; `setValue`: (`v`: `string` \| `number`) => `void`  }) => `Element` |
-| `ConfigElementTime` | (`props`: { `disabled?`: `boolean` ; `oldValue`: { `hour`: `number` ; `minute`: `number` ; `second`: `number`  } ; `title`: `string` ; `value`: { `hour`: `number` ; `minute`: `number` ; `second`: `number`  } ; `setValue`: (`v`: { `hour`: `number` ; `minute`: `number` ; `second`: `number`  }) => `void`  }) => `Element` |
-| `ConfigElementToggle` | (`props`: { `oldValue`: `boolean` ; `title`: `string` ; `value`: `boolean` ; `setValue`: (`v`: `boolean`) => `void`  }) => `Element` |
-| `ConfigSectionCamtracker` | () => `Element` |
-| `ConfigSectionErrorEmail` | () => `Element` |
-| `ConfigSectionGeneral` | () => `Element` |
-| `ConfigSectionHelios` | () => `Element` |
-| `ConfigSectionMeasurementTriggers` | () => `Element` |
-| `ConfigSectionOpus` | () => `Element` |
-| `ConfigSectionTumPlc` | () => `Element` |
-| `ConfigSectionUpload` | () => `Element` |
-| `LabeledRow` | (`props`: { `children`: `ReactNode` ; `modified?`: `boolean` ; `title`: `string`  }) => `Element` |
-| `SavingOverlay` | (`props`: { `errorMessage`: `undefined` \| `string` ; `isSaving`: `boolean` ; `resetLocalConfig`: () => `void` ; `saveLocalConfig`: () => `void`  }) => `Element` |
-
-#### Defined in
-
-[components/configuration/index.ts:15](https://github.com/tum-esm/pyra/blob/8f435c2/packages/ui/src/components/configuration/index.ts#L15)
-
-___
-
-### essentialComponents
-
-• **essentialComponents**: `Object`
-
-#### Type declaration
-
-| Name | Type |
-| :------ | :------ |
-| `Button` | (`props`: { `children`: `ReactNode` ; `className?`: `string` ; `disabled?`: `boolean` ; `dot?`: `boolean` ; `spinner?`: `boolean` ; `variant`: ``"white"`` \| ``"green"`` \| ``"red"`` \| ``"gray"`` \| ``"flat-blue"`` ; `onClick`: () => `void`  }) => `Element` |
-| `LiveSwitch` | (`props`: { `isLive`: `boolean` ; `toggle`: (`v`: `boolean`) => `void`  }) => `Element` |
[…]
