+++ packages/docs/docs/api-reference/ui/index.md
@@ -1,16 +1,0 @@
----
-id: "index"
-title: "pyra-ui"
-sidebar_label: "Table of Contents"
-sidebar_position: 0.5
-hide_table_of_contents: true
-custom_edit_url: null
----
-
-## Modules
-
-- [assets](modules/assets.md)
-- [components](modules/components.md)
-- [custom-types](modules/custom_types.md)
-- [tabs](modules/tabs.md)
-- [utils](modules/utils.md)
