+++ packages/docs/docs/api-reference/ui/_category_.yml
@@ -1,2 +1,0 @@
-label: "ui"
-position: 3
