+++ packages/docs/docs/api-reference/ui/modules/utils.md
@@ -1,76 +1,0 @@
----
-id: "utils"
-title: "Module: utils"
-sidebar_label: "utils"
-sidebar_position: 0
-custom_edit_url: null
----
-
-## Variables
-
-### fetchUtils
-
-• **fetchUtils**: `Object`
-
-#### Type declaration
-
-| Name | Type |
-| :------ | :------ |
-| `backend` | { `archiveLogs`: () => `Promise`<`ChildProcess`\> ; `checkPyraCoreState`: () => `Promise`<`ChildProcess`\> ; `getConfig`: () => `Promise`<`ChildProcess`\> ; `pyraCliIsAvailable`: () => `Promise`<`ChildProcess`\> ; `startPyraCore`: () => `Promise`<`ChildProcess`\> ; `stopPyraCore`: () => `Promise`<`ChildProcess`\> ; `updateConfig`: (`newConfig`: [`partialConfig`](../namespaces/custom_types.customTypes.md#partialconfig)) => `Promise`<`ChildProcess`\> ; `writeToPLC`: (`command`: `string`[]) => `Promise`<`ChildProcess`\>  } |
-| `backend.archiveLogs` | () => `Promise`<`ChildProcess`\> |
-| `backend.checkPyraCoreState` | () => `Promise`<`ChildProcess`\> |
-| `backend.getConfig` | () => `Promise`<`ChildProcess`\> |
-| `backend.pyraCliIsAvailable` | () => `Promise`<`ChildProcess`\> |
-| `backend.startPyraCore` | () => `Promise`<`ChildProcess`\> |
-| `backend.stopPyraCore` | () => `Promise`<`ChildProcess`\> |
-| `backend.updateConfig` | (`newConfig`: [`partialConfig`](../namespaces/custom_types.customTypes.md#partialconfig)) => `Promise`<`ChildProcess`\> |
-| `backend.writeToPLC` | (`command`: `string`[]) => `Promise`<`ChildProcess`\> |
-| `getFileContent` | (`filePath`: `string`) => `Promise`<`string`\> |
-| `getProjectDirPath` | () => `Promise`<`string`\> |
-| `initialAppState` | (`dispatch`: `any`, `setBackendIntegrity`: (`value`: `undefined` \| ``"cli is missing"`` \| ``"config is invalid"`` \| ``"pyra-core is not running"`` \| ``"valid"``) => `void`) => `Promise`<`void`\> |
-
-#### Defined in
-
-[utils/fetch-utils/index.ts:6](https://github.com/tum-esm/pyra/blob/8f435c2/packages/ui/src/utils/fetch-utils/index.ts#L6)
-
-___
-
-### functionalUtils
-
-• **functionalUtils**: `Object`
-
-#### Type declaration
-
-| Name | Type |
-| :------ | :------ |
-| `deepEqual` | (`a`: `any`, `b`: `any`) => `boolean` |
-| `generateActivityHistories` | (`activityHistory`: [`activityHistory`](../namespaces/custom_types.customTypes.md#activityhistory), `measurementsAreRunning`: `boolean`, `errorIsPresent`: `boolean`) => { `core`: `ActivitySection`[] ; `error`: `ActivitySection`[] ; `measurements`: `ActivitySection`[]  } |
-| `parseNumberTypes` | (`newConfig`: [`config`](../namespaces/custom_types.customTypes.md#config)) => [`config`](../namespaces/custom_types.customTypes.md#config) |
-| `reduceLogLines` | (`logLines`: `string`[], `mode`: ``"3 iterations"`` \| ``"5 minutes"``) => `string`[] |
-
-#### Defined in
-
[…]
