+++ packages/docs/docs/api-reference/cli/_category_.yml
@@ -1,2 +1,0 @@
-label: 'cli'
-position: 2
