+++ packages/docs/docs/api-reference/core/_category_.yml
@@ -1,2 +1,0 @@
-label: 'core'
-position: 1
