### 2023-03-28 035e9a4
Add test of API reference state
