+++ pyproject.toml
@@ -24,9 +24,9 @@ pydantic = "^1.9.2"
 tqdm = "^4.64.0"
 colorama = "^0.4.5"
 deepdiff = "^6.2.1"
+tum-esm-utils = "^1.3.0"
 
 # libraries only used in development
-tum-esm-utils = "^1.1.0"
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.1.2"
 black = "^22.3.0"
