### 2023-03-28 721a0c1
Check static types on tests
