+++ tests/integration/test_camtracker_connection.py
@@ -3,6 +3,6 @@ from packages.core import interfaces, modules
 
 
 @pytest.mark.integration
-def test_opus_connection():
+def test_opus_connection() -> None:
     config = interfaces.ConfigInterface.read()
     modules.sun_tracking.SunTracking(config).test_setup()
