### 2023-03-28 b2bed9c
Check static types on tests
