+++ scripts/check_static_types.sh
@@ -10,3 +10,6 @@ python -m mypy run_pyra_core.py
 
 echo "Checking packages/cli/main.py"
 python -m mypy packages/cli/main.py
+
+echo "Checking tests/"
+python -m mypy tests/
