+++ scripts/check_static_types.sh
@@ -6,10 +6,10 @@ echo "Removing old mypy cache"
 rm -rf .mypy_cache 
 
 echo "Checking run_pyra_core.py"
-python -m mypy run_pyra_core.py
+mypy run_pyra_core.py
 
 echo "Checking packages/cli/main.py"
-python -m mypy packages/cli/main.py
+mypy packages/cli/main.py
 
 echo "Checking tests/"
-python -m mypy tests/
+mypy tests/
