### 2023-03-28 724195e
Clean up CI workflow file
