+++ .github/workflows/test-python-codebase-on-main.yml
@@ -59,19 +59,12 @@ jobs:
               run: |
                   python -m venv .venv
                   source .venv/bin/activate
-                  which python
-                  python --version
                   poetry install --with=dev
 
             # run test suite
             - name: Run mypy static type analysis
               run: |
-                  ls -lahg .
-                  which python
-                  pip list
                   source .venv/bin/activate
-                  which python
-                  pip list
                   bash scripts/check_static_types.sh
             - name: Run pytest tests
               run: |
