+++ .github/workflows/test-python-codebase-on-main.yml
@@ -31,43 +31,36 @@ jobs:
             # check-out repo
             - name: Check out repository
               uses: actions/checkout@v3
-
-            # load cached venv if cache exists
-            - name: Load cached venv
-              id: cached-poetry-dependencies
-              uses: actions/cache@v3
-              with:
-                  path: .venv
-                  key: venv-${{ runner.os }}-3.10.6-${{ hashFiles('poetry.lock') }}
-
-            # set up venv if not cached
             - name: Set up Python 3.10.6
-              if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
-              uses: actions/setup-python@v3
+              uses: actions/setup-python@v4
               with:
                   python-version: 3.10.6
-            - name: Install Poetry
-              if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
+
+            - name: Install and configure Poetry
               uses: snok/install-poetry@v1
               with:
                   version: 1.3.2
                   virtualenvs-create: true
                   virtualenvs-in-project: true
                   installer-parallel: true
+
+            - name: Load cached venv
+              id: cached-poetry-dependencies
+              uses: actions/cache@v3
+              with:
+                  path: ~/.cache
+                  key: venv-${{ runner.os }}-3.10.6-${{ hashFiles('poetry.lock') }}
+
             - name: Install dependencies
-              if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
-              run: |
-                  python -m venv .venv
-                  source .venv/bin/activate
-                  poetry install --with=dev
+              run: poetry install --no-interaction --no-root
 
             # run test suite
             - name: Run mypy static type analysis
               run: |
-                  source .venv/bin/activate
+                  source $VENV
                   bash scripts/check_static_types.sh
             - name: Run pytest tests
               run: |
-                  source .venv/bin/activate
+                  source $VENV
                   pytest -m "ci" --cov=packages tests
                   coverage report
