+++ .github/workflows/test-python-codebase-on-main.yml
@@ -61,10 +61,10 @@ jobs:
             # run test suite
             - name: Run mypy static type analysis
               run: |
-                  source .venv/bin/activated
+                  source .venv/bin/activate
                   bash scripts/check_static_types.sh
             - name: Run pytest tests
               run: |
-                  source .venv/bin/activated
+                  source .venv/bin/activate
                   pytest -m "ci" --cov=packages tests 
                   coverage report
