+++ .github/workflows/test-python-codebase-on-main.yml
@@ -28,43 +28,28 @@ jobs:
     test:
         runs-on: ubuntu-latest
         steps:
-            # check-out repo
-            - name: Check out repository
+            #-------------------------------------#
+            #  Check out repo and set up Python   #
+            #-------------------------------------#
+            - name: Check out the repository
               uses: actions/checkout@v3
-
-            # load cached venv if cache exists
-            - name: Load cached venv
-              id: cached-poetry-dependencies
-              uses: actions/cache@v3
-              with:
-                  path: .venv
-                  key: venv-${{ runner.os }}-3.10.6-${{ hashFiles('poetry.lock') }}
-
-            # generate venv if not in cache
-            - name: Set up Python 3.10.6
-              if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
-              uses: actions/setup-python@v4
+            - name: 'Setup Python, Poetry and Dependencies'
+              uses: packetcoders/action-setup-cache-python-poetry@main
               with:
                   python-version: 3.10.6
-            - name: Install Poetry
-              if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
-              uses: snok/install-poetry@v1
-              with:
-                  version: 1.3.2
-                  virtualenvs-create: true
-                  virtualenvs-in-project: true
-                  installer-parallel: true
-            - name: Install dependencies
-              if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
-              run: poetry install --with=dev
+                  poetry-version: 1.3.2
+
+            #------------------------#
+            #  run test suite        #
+            #------------------------#
+            - name: Run tests
+              run: |
+                  poetry run pytest
 
             # run test suite
             - name: Run mypy static type analysis
-              run: |
-                  source .venv/bin/activate
-                  bash scripts/check_static_types.sh
+              run: bash scripts/check_static_types.sh
             - name: Run pytests
               run: |
-                  source .venv/bin/activate
                   pytest -m "ci" --cov=packages tests 
                   coverage report
