+++ .github/workflows/test-python-codebase-on-main.yml
@@ -42,11 +42,6 @@ jobs:
             #------------------------#
             #  run test suite        #
             #------------------------#
-            - name: Run tests
-              run: |
-                  poetry run pytest
-
-            # run test suite
             - name: Run mypy static type analysis
               run: bash scripts/check_static_types.sh
             - name: Run pytests
