+++ .github/workflows/test-python-codebase-on-main.yml
@@ -28,23 +28,45 @@ jobs:
     test:
         runs-on: ubuntu-latest
         steps:
-            #-------------------------------------#
-            #  Check out repo and set up Python   #
-            #-------------------------------------#
-            - name: Check out the repository
+            # check-out repo and set-up python
+            - name: Check out repository
               uses: actions/checkout@v3
-            - name: 'Setup Python, Poetry and Dependencies'
-              uses: packetcoders/action-setup-cache-python-poetry@main
+
+            # load cached venv if cache exists
+            # - name: Load cached venv
+            #   id: cached-poetry-dependencies
+            #   uses: actions/cache@v3
+            #   with:
+            #       path: .venv
+            #       key: venv-${{ runner.os }}-3.10.6-${{ hashFiles('poetry.lock') }}
+
+            # set up venv if not cached
+            - name: Set up Python 3.10.6
+              #if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
+              uses: actions/setup-python@v3
               with:
                   python-version: 3.10.6
-                  poetry-version: 1.3.2
+            - name: Install Poetry
+              #if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
+              uses: snok/install-poetry@v1
+              with:
+                  version: 1.3.2
+                  virtualenvs-create: true
+                  virtualenvs-in-project: true
+                  installer-parallel: true
 
-            #------------------------#
-            #  run test suite        #
-            #------------------------#
+            # install dependencies if cache does not exist
+            - name: Install dependencies
+              #if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
+              run: poetry install
+
+            # run test suite
             - name: Run mypy static type analysis
-              run: bash scripts/check_static_types.sh
-            - name: Run pytests
               run: |
-                  pytest -m "ci" --cov=packages tests 
+                  source .venv/bin/activate
+                  bash scripts/check_static_types.sh
+            - name: Run pytest tests
+              run: |
+                  source .venv/bin/activate
+                  pytest -m "ci" --cov=packages tests
                   coverage report
