+++ .github/workflows/test-python-codebase-on-main.yml
@@ -46,8 +46,21 @@ jobs:
               uses: actions/setup-python@v3
               with:
                   python-version: 3.10.6
+
+            - name: Load cached venv
+              id: cached-poetry-dependencies
+              uses: actions/cache@v3
+              with:
+                  path: .venv
+                  key: venv-${{ runner.os }}-${{ hashFiles('**/poetry.lock') }}
+
+            - name: Load cached .local
+              uses: actions/cache@v3
+              with:
+                  path: ~/.local
+                  key: dotlocal-${{ runner.os }}-${{ hashFiles('.github/workflows/test-python-codebase-on-main.yml') }}
+
             - name: Install Poetry
-              #if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
               uses: snok/install-poetry@v1
               with:
                   version: 1.3.2
@@ -57,7 +70,7 @@ jobs:
 
             # install dependencies if cache does not exist
             - name: Install dependencies
-              #if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
+              if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
               run: poetry install
 
             # run test suite
