+++ .github/workflows/test-python-codebase-on-main.yml
@@ -31,35 +31,25 @@ jobs:
             # check-out repo and set-up python
             - name: Check out repository
               uses: actions/checkout@v3
-
-            # load cached venv if cache exists
-            # - name: Load cached venv
-            #   id: cached-poetry-dependencies
-            #   uses: actions/cache@v3
-            #   with:
-            #       path: .venv
-            #       key: venv-${{ runner.os }}-3.10.6-${{ hashFiles('poetry.lock') }}
-
-            # set up venv if not cached
             - name: Set up Python 3.10.6
-              #if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
-              uses: actions/setup-python@v3
+              uses: actions/setup-python@v4
               with:
                   python-version: 3.10.6
 
+            # load caches (dependencies and poetry binaries)
             - name: Load cached venv
               id: cached-poetry-dependencies
               uses: actions/cache@v3
               with:
                   path: .venv
                   key: venv-${{ runner.os }}-${{ hashFiles('**/poetry.lock') }}
-
             - name: Load cached .local
               uses: actions/cache@v3
               with:
                   path: ~/.local
                   key: dotlocal-${{ runner.os }}-${{ hashFiles('.github/workflows/test-python-codebase-on-main.yml') }}
 
+            # install poetry (will be quick after it has been cached)
             - name: Install Poetry
               uses: snok/install-poetry@v1
               with:
