+++ packages/core/utils/update_dict_recursively.py
@@ -1,37 +1,12 @@
 from typing import Any
 
 
+# TODO: use tum_esm_utils
+
+
 def update_dict_recursively(old_object: Any, new_object: Any) -> Any:
-    """
-    For a given dict, update it recursively from a new dict.
-    It will not add any properties and assert that the types
-    remain the same (or null). null->int or int->null is possible
-    but not int->dict or list->int.
-
-    example:
-
-    ```python
-    old_object = {
-        "a": 3,
-        "b": {
-            "c": 50,
-            "e": null
-        }
-    }
-    new_object = {
-        "b": {
-            "e": 80
-        }
-    }
-    returned_object = {
-        "a": 3,
-        "b": {
-            "c": 50,
-            "e": 80
-        }
-    }
-    ```
-    """
+    """TO BE REMOVED IN 4.0.8: Pyra will use `merge_dicts` from
+    `tum_esm_utils` instead."""
 
     if old_object is None or new_object is None:
         return new_object
