+++ packages/core/utils/with_filelock.py
@@ -6,16 +6,12 @@ from functools import wraps
 # https://github.com/python/mypy/issues/1551#issuecomment-253978622
 F = TypeVar("F", bound=Callable[..., Any])
 
+# TODO: use tum_esm_utils
+
 
 class with_filelock:
-    """
-    FileLock = Mark, that a file is being used and other programs
-    should not interfere. A file "*.lock" will be created and the
-    content of this file will make the wrapped function possibly
-    wait until other programs are done using it.
-
-    See https://en.wikipedia.org/wiki/Semaphore_(programming)
-    """
+    """TO BE REMOVED IN 4.0.8: Pyra will use `with_filelock` from
+    `tum_esm_utils` instead."""
 
     def __init__(self, file_lock_path: str, timeout: float = -1) -> None:
         """A timeout of -1 means that the code waits forever"""
