+++ packages/cli/commands/remove_filelocks.py
@@ -1,3 +1,6 @@
+"""Remove all filelocks. Helpful when any of the programs crashed during
+writing to a file. Normally, this should not be necessary."""
+
 import click
 import os
 
