+++ packages/core/utils/update_dict_recursively.py
@@ -5,8 +5,20 @@ from typing import Any
 
 
 def update_dict_recursively(old_object: Any, new_object: Any) -> Any:
-    """TO BE REMOVED IN 4.0.8: Pyra will use `merge_dicts` from
-    `tum_esm_utils` instead."""
+    """Update a given dict recursively with data from a new dict.
+
+    It will not add any properties and assert that the types
+    remain the same (or null). null->int or int->null is possible
+    but not int->dict or list->int.
+
+    example:
+    ```python
+    merge_dicts(
+        old_object={"a": 3, "b": {"c": 50, "e": None}},
+        new_object={"b": {"e": 80}},
+    ) == {"a": 3, "b": {"c": 50, "e": 80}}
+    ```
+    """
 
     if old_object is None or new_object is None:
         return new_object
