### 2023-04-03 0c4ad96
Add docstrings to TypeDicts
