+++ packages/core/types/config.py
@@ -4,20 +4,48 @@ from typing import Any, Callable, Literal, Optional, TypedDict
 
 
 class TimeDict(TypedDict):
+    """TypedDict:
+
+    ```ts
+    {
+        hour: int;
+        minute: int;
+        second: int;
+    }
+    ```
+    """
+
     hour: int
     minute: int
     second: int
 
 
 class TimeDictPartial(TypedDict, total=False):
+    """TypedDict: like `TimeDict`, but all fields are optional."""
+
     hour: int
     minute: int
     second: int
 
 
 class ConfigSubDicts:
+    """Class that contains TypedDicts for sections of the config file."""
+
     @staticmethod
     class General(TypedDict):
+        """TypedDict:
+
+        ```ts
+        {
+            version: "4.0.7";
+            seconds_per_core_interval: float;
+            test_mode: bool;
+            station_id: str;
+            min_sun_elevation: float;
+        }
+        ```
+        """
+
         version: Literal["4.0.7"]
         seconds_per_core_interval: float
         test_mode: bool
@@ -26,6 +54,9 @@ class ConfigSubDicts:
 
     @staticmethod
     class GeneralPartial(TypedDict, total=False):
+        """TypedDict: like `ConfigSubDicts.General`, but all fields are
+        optional."""
+
         seconds_per_core_interval: float
         test_mode: bool
         station_id: str
@@ -33,6 +64,20 @@ class ConfigSubDicts:
 
     @staticmethod
     class Opus(TypedDict):
+        """TypedDict:
+
+        ```ts
+        {
+            em27_ip: str;
+            executable_path: str;
+            experiment_path: str;
+            macro_path: str;
+            username: str;
+            password: str;
+        }
+        ```
+        """
+
         em27_ip: str
         executable_path: str
         experiment_path: str
@@ -42,6 +87,9 @@ class ConfigSubDicts:
 
     @staticmethod
     class OpusPartial(TypedDict, total=False):
+        """TypedDict: like `ConfigSubDicts.Opus`, but all fields are
+        optional."""
+
         em27_ip: str
         executable_path: str
         experiment_path: str
@@ -51,6 +99,19 @@ class ConfigSubDicts:
 
     @staticmethod
     class Camtracker(TypedDict):
+        """TypedDict:
+
+        ```ts
+        {
+            config_path: str;
+            executable_path: str;
+            learn_az_elev_path: str;
+            sun_intensity_path: str;
+            motor_offset_threshold: float;
+        }
+        ```
+        """
+
         config_path: str
         executable_path: str
         learn_az_elev_path: str
@@ -59,6 +120,9 @@ class ConfigSubDicts:
 
     @staticmethod
     class CamtrackerPartial(TypedDict, total=False):
+        """TypedDict: like `ConfigSubDicts.Camtracker` but all fields
+        are optional."""
+
         config_path: str
         executable_path: str
         learn_az_elev_path: str
@@ -67,6 +131,18 @@ class ConfigSubDicts:
 
     @staticmethod
     class ErrorEmail(TypedDict):
+        """TypedDict:
+
+        ```ts
+        {
+            sender_address: str;
+            sender_password: str;
+            notify_recipients: bool;
+            recipients: str;
+        }
+        ```
+        """
+
         sender_address: str
         sender_password: str
         notify_recipients: bool
@@ -74,6 +150,9 @@ class ConfigSubDicts:
 
     @staticmethod
     class ErrorEmailPartial(TypedDict, total=False):
+        """TypedDict: like `ConfigSubDicts.ErrorEmail` but all fields
+        are optional."""
+
         sender_address: str
         sender_password: str
         notify_recipients: bool
@@ -81,18 +160,46 @@ class ConfigSubDicts:
 
     @staticmethod
[…]
