### 2023-04-03 3b39986
Add docstrings to TypeDicts
