+++ joss/paper.bib
@@ -1,342 +1,418 @@
 @article{jones2020assessing,
-  author  = {Jones, T. S. and Franklin, J. E. and Chen, J. and Dietrich, F. and Hajny, K. D. and Paetzold, J. C. and Wenzel, A. and Gately, C. and Gottlieb, E. and Parker, H. and Dubey, M. and Hase, F. and Shepson, P. B. and Mielke, L. H. and Wofsy, S. C.},
-  title   = {Assessing urban methane emissions using column-observing portable Fourier transform infrared (FTIR) spectrometers and a novel Bayesian inversion framework},
-  journal = {Atmospheric Chemistry and Physics},
-  volume  = {21},
-  year    = {2021},
-  number  = {17},
-  pages   = {13131--13147},
-  url     = {https://acp.copernicus.org/articles/21/13131/2021/},
-  doi     = {10.5194/acp-21-13131-2021}
+  author     = {Jones, T. S. and Franklin, J. E. and Chen, J. and Dietrich, F. and Hajny, K. D. and Paetzold, J. C. and Wenzel, A. and Gately, C. and Gottlieb, E. and Parker, H. and Dubey, M. and Hase, F. and Shepson, P. B. and Mielke, L. H. and Wofsy, S. C.},
+  doi        = {10.5194/acp-21-13131-2021},
+  journal    = {Atmospheric Chemistry and Physics},
+  number     = {17},
+  pages      = {13131--13147},
+  title      = {Assessing urban methane emissions using column-observing portable Fourier transform infrared (FTIR) spectrometers and a novel Bayesian inversion framework},
+  url        = {https://acp.copernicus.org/articles/21/13131/2021/},
+  volume     = {21},
+  year       = {2021},
+  bdsk-url-1 = {https://acp.copernicus.org/articles/21/13131/2021/},
+  bdsk-url-2 = {https://doi.org/10.5194/acp-21-13131-2021}
 }
 
 @article{dietrich2021muccnet,
-  author  = {Dietrich, F. and Chen, J. and Voggenreiter, B. and Aigner, P. and Nachtigall, N. and Reger, B.},
-  title   = {MUCCnet: Munich Urban Carbon Column network},
-  journal = {Atmospheric Measurement Techniques},
-  volume  = {14},
-  year    = {2021},
-  number  = {2},
-  pages   = {1111--1126},
-  url     = {https://amt.copernicus.org/articles/14/1111/2021/},
-  doi     = {10.5194/amt-14-1111-2021}
+  author     = {Dietrich, F. and Chen, J. and Voggenreiter, B. and Aigner, P. and Nachtigall, N. and Reger, B.},
+  doi        = {10.5194/amt-14-1111-2021},
+  journal    = {Atmospheric Measurement Techniques},
+  number     = {2},
+  pages      = {1111--1126},
+  title      = {MUCCnet: Munich Urban Carbon Column network},
+  url        = {https://amt.copernicus.org/articles/14/1111/2021/},
+  volume     = {14},
+  year       = {2021},
+  bdsk-url-1 = {https://amt.copernicus.org/articles/14/1111/2021/},
+  bdsk-url-2 = {https://doi.org/10.5194/amt-14-1111-2021}
 }
 
 
 @article{chen2020oktoberfest,
-  author  = {Chen, J. and Dietrich, F. and Maazallahi, H. and Forstmaier, A. and Winkler, D. and Hofmann, M. E. G. and Denier van der Gon, H. and R\"ockmann, T.},
-  title   = {Methane emissions from the Munich Oktoberfest},
-  journal = {Atmospheric Chemistry and Physics},
-  volume  = {20},
-  year    = {2020},
-  number  = {6},
-  pages   = {3683--3696},
-  url     = {https://acp.copernicus.org/articles/20/3683/2020/},
[…]
