+++ joss/paper.bib
@@ -421,7 +421,7 @@
   editor    = {Masson-Delmotte, V. and Zhai, P. and Pirani, A. and Connors, S.L. and Péan, C. and Berger, S. and Caud, N. and Chen, Y. and Goldfarb, L. and Gomis, M.I. and Huang, M. and Leitzell, K. and Lonnoy, E. and Matthews, J.B.R. and Maycock, T.K. and Waterfield, T. and Yelekçi, O. and Yu, R. and Zhou, B.},
   publisher = {Cambridge University Press},
   address   = {Cambridge, United Kingdom and New York, NY, USA},
-  pages     = {33-144},
+  pages     = {33--144},
   doi       = {10.1017/9781009157896.002},
   year      = {2021},
   type      = {Book Section}
@@ -447,7 +447,7 @@
   title     = {Novel methane emission estimation method for ground based remote sensing networks},
   author    = {Klappenbach, F., Chen, J., Wenzel, A., Dietrich, F., Forstermeier, A., Zhao, X., Jones, T., Franklin, J., Wofsy, S., Frey, M., Hase, F., Hedelius, J., Wennberg, P., and Cohen, R.},
   booktitle = {EGU General Assembly 2022},
-  pages     = {EGU22-10604},
+  pages     = {EGU22--10604},
   year      = {2022},
   doi       = {10.5194/egusphere-egu22-10604},
   url       = {https://doi.org/10.5194/egusphere-egu22-10604}
