### 2023-04-04 fef7840
Add GHG and FTIR acronym introduction #158
