+++ joss/paper.md
@@ -57,7 +57,7 @@ bibliography:
 
 # Statement of need
 
-The human activity of burning fossil fuels is the driving factor of global warming [@IPCC2021]. To promote global climate change mitigation policies, a profound GHG observational data basis is required to understand and quantify the sources and sinks of GHG emissions [@IPCC2021]. Ground-based remote sensing instruments that analyze direct sunlight using, for example, EM27/SUN Fourier transform spectrometers [@gisi2011camtracker] fill the gap between ground-based in situ measurements and space-based measurements by satellites [@hase2015application;@rissmann2022gradients]. Due to its dependency on direct sunlight and clear skies, the EM27/SUN generally requires a trained operator on site. Depending on the weather conditions, the operator needs to manually control the instrument and measurement times. This is time and cost-intensive, in particular, if more than one instrument is involved. However, state-of-the-art EM27/SUN networks consist of up to 6 instruments to estimate the emissions of cities [@jones2020assessing;@dietrich2021muccnet;@tu2022madridwastedisposal;@hase2015application;@ionov2021stpetersburg;@che2022lagrangian;@vogel2019paris] or local and regional GHG sources [@chen2020oktoberfest;@luther2022observational;@tu2020intercomparison;@toja2017cfdco2dispersion;@chen2016differentialcolumn;@tu2022madridwastedisposal;@klappenbach2022sanfran]. These setups generally require at least one operator to operate and closely monitor each instrument.
+The human activity of burning fossil fuels is the driving factor of global warming [@IPCC2021]. To promote global climate change mitigation policies, a profound greenhouse gases (GHG) observational data basis is required to understand and quantify the sources and sinks of GHG emissions [@IPCC2021]. Ground-based remote sensing instruments that analyze direct sunlight using, for example, EM27/SUN fourier-transform infrared spectrometers (FTIR) [@gisi2011camtracker] fill the gap between ground-based in situ measurements and space-based measurements by satellites [@hase2015application;@rissmann2022gradients]. Due to its dependency on direct sunlight and clear skies, the EM27/SUN generally requires a trained operator on site. Depending on the weather conditions, the operator needs to manually control the instrument and measurement times. This is time and cost-intensive, in particular, if more than one instrument is involved. However, state-of-the-art EM27/SUN networks consist of up to 6 instruments to estimate the emissions of cities [@jones2020assessing;@dietrich2021muccnet;@tu2022madridwastedisposal;@hase2015application;@ionov2021stpetersburg;@che2022lagrangian;@vogel2019paris] or local and regional GHG sources [@chen2020oktoberfest;@luther2022observational;@tu2020intercomparison;@toja2017cfdco2dispersion;@chen2016differentialcolumn;@tu2022madridwastedisposal;@klappenbach2022sanfran]. These setups generally require at least one operator to operate and closely monitor each instrument.
 _Pyra_ is an automation software that does not require a trained operator and enables the user to perform EM27/SUN measurements with a workload limited to system monitoring. This also guarantees continuous measurements during weekends, holidays, and at remote locations resulting in an optimized data yield while saving time and human resources. _Pyra_'s open-source code and documentation will reduce operator training efforts.
 
[…]
