### 2023-04-04 0f33ea1
Fix author linking in EGU references
