+++ joss/paper.bib
@@ -390,7 +390,7 @@
 
 @inproceedings{humpage2021uganda,
   title     = {Greenhouse gas column observations from a portable spectrometer in Uganda},
-  author    = {Humpage, N., Boesch, H., Okello, W., Dietrich, F., Chen, J., Lunt, M., Feng, L., Palmer, P., and Hase, F.},
+  author    = {Humpage, N. and Boesch, H. and Okello, W. and Dietrich, F. and Chen, J. and Lunt, M. and Feng, L. and Palmer, P. and Hase, F.},
   booktitle = {EGU General Assembly 2021},
   pages     = {EGU21--10156},
   year      = {2021},
@@ -445,7 +445,7 @@
 
 @inproceedings{klappenbach2022sanfran,
   title     = {Novel methane emission estimation method for ground based remote sensing networks},
-  author    = {Klappenbach, F., Chen, J., Wenzel, A., Dietrich, F., Forstermeier, A., Zhao, X., Jones, T., Franklin, J., Wofsy, S., Frey, M., Hase, F., Hedelius, J., Wennberg, P., and Cohen, R.},
+  author    = {Klappenbach, F., and Chen, J., and Wenzel, A., and Dietrich, F., and Forstermeier, A., and Zhao, X., and Jones, T., and Franklin, J., and Wofsy, S., and Frey, M., and Hase, F., and Hedelius, J., and Wennberg, P., and Cohen, R.},
   booktitle = {EGU General Assembly 2022},
   pages     = {EGU22--10604},
   year      = {2022},
