### 2023-04-04 fe9a0ae
Remove \chem stuff in LaTeX
