+++ joss/paper.bib
@@ -124,7 +124,7 @@
   journal    = {Atmospheric Measurement Techniques},
   number     = {9},
   pages      = {4751--4771},
-  title      = {Intercomparison of atmospheric \chem{CO_2} and \chem{CH_4} abundances on regional scales in boreal areas using Copernicus Atmosphere Monitoring Service (CAMS) analysis, COllaborative Carbon Column Observing Network (COCCON) spectrometers, and Sentinel-5 Precursor satellite observations},
+  title      = {Intercomparison of atmospheric CO2 and CH4 abundances on regional scales in boreal areas using Copernicus Atmosphere Monitoring Service (CAMS) analysis, COllaborative Carbon Column Observing Network (COCCON) spectrometers, and Sentinel-5 Precursor satellite observations},
   url        = {https://amt.copernicus.org/articles/13/4751/2020/},
   volume     = {13},
   year       = {2020},
@@ -212,7 +212,7 @@
   journal    = {Atmospheric Measurement Techniques},
   number     = {10},
   pages      = {5217--5230},
-  title      = {Quantifying \chem{CH_4} emissions from hard coal mines using mobile sun-viewing Fourier transform spectrometry},
+  title      = {Quantifying CH4 emissions from hard coal mines using mobile sun-viewing Fourier transform spectrometry},
   url        = {https://amt.copernicus.org/articles/12/5217/2019/},
   volume     = {12},
   year       = {2019},
@@ -242,7 +242,7 @@
   journal    = {Earth System Science Data},
   number     = {1},
   pages      = {199--211},
-  title      = {Shipborne measurements of \chem{XCO_2}, \chem{XCH_4}, and \chem{XCO} above the Pacific Ocean and comparison to CAMS atmospheric analyses and S5P/TROPOMI},
+  title      = {Shipborne measurements of XCO2, XCH4, and XCO above the Pacific Ocean and comparison to CAMS atmospheric analyses and S5P/TROPOMI},
   url        = {https://essd.copernicus.org/articles/13/199/2021/},
   volume     = {13},
   year       = {2021},
@@ -257,7 +257,7 @@
   journal    = {Atmospheric Chemistry and Physics},
   number     = {5},
   pages      = {3271--3285},
-  title      = {\chem{XCO_2} in an emission hot-spot region: the COCCON Paris campaign 2015},
+  title      = {XCO2 in an emission hot-spot region: the COCCON Paris campaign 2015},
   url        = {https://acp.copernicus.org/articles/19/3271/2019/},
   volume     = {19},
   year       = {2019},
@@ -365,7 +365,7 @@
   journal    = {Earth System Science Data},
   number     = {3},
   pages      = {935--946},
-  title      = {Satellite and ground-based measurements of \chem{XCO_2} in a remote semiarid region of Australia},
+  title      = {Satellite and ground-based measurements of XCO2 in a remote semiarid region of Australia},
   url        = {https://essd.copernicus.org/articles/11/935/2019/},
   volume     = {11},
   year       = {2019},
