### 2023-04-04 7956645
Remove \mathrm stuff in LaTeX
