+++ joss/paper.bib
@@ -318,7 +318,7 @@
   journal    = {Atmospheric Measurement Techniques},
   number     = {8},
   pages      = {3527--3546},
-  title      = {Assessment of errors and biases in retrievals of X$_\mathrm{CO_2}$, X$_\mathrm{CH_4}$, X$_\mathrm{CO}$, and X$_\mathrm{N_{2}O}$ from a 0.5\,cm$^{-1}$ resolution solar-viewing spectrometer},
+  title      = {Assessment of errors and biases in retrievals of XCO2, XCH4, XCO, and XN2O from a 0.5\,cm$^{-1}$ resolution solar-viewing spectrometer},
   url        = {https://amt.copernicus.org/articles/9/3527/2016/},
   volume     = {9},
   year       = {2016},
@@ -434,7 +434,7 @@
   journal    = {Atmospheric Measurement Techniques},
   number     = {22},
   pages      = {6605--6623},
-  title      = {Comparison of OCO-2 target observations to MUCCnet -- is it possible to capture urban $X_\mathrm{CO_2}$ gradients from space?},
+  title      = {Comparison of OCO-2 target observations to MUCCnet -- is it possible to capture urban XCO2 gradients from space?},
   url        = {https://amt.copernicus.org/articles/15/6605/2022/},
   volume     = {15},
   year       = {2022},
