+++ joss/paper.bib
@@ -287,7 +287,7 @@
   journal    = {Atmospheric Chemistry and Physics},
   number     = {13},
   pages      = {8479--8498},
-  title      = {Differential column measurements using compact \hack{\newline}solar-tracking spectrometers},
+  title      = {Differential column measurements using compact solar-tracking spectrometers},
   url        = {https://acp.copernicus.org/articles/16/8479/2016/},
   volume     = {16},
   year       = {2016},
