### 2023-04-04 4eddd75
Improve bibliography rendering #158
