+++ joss/paper.bib
@@ -63,7 +63,7 @@
   journal    = {Atmospheric Chemistry and Physics},
   number     = {1},
   pages      = {295--317},
-  title      = {Quantification of CH$_{4}$ emissions from waste disposal sites near the city of Madrid using ground- and space-based observations of COCCON, IASI},
+  title      = {Quantification of $CH_4$ emissions from waste disposal sites near the city of Madrid using ground- and space-based observations of COCCON, IASI},
   url        = {https://acp.copernicus.org/articles/22/295/2022/},
   volume     = {22},
   year       = {2022},
@@ -95,7 +95,7 @@
   number     = {7},
   pages      = {075001},
   publisher  = {IOP Publishing},
-  title      = {Lagrangian inversion of anthropogenic CO2 emissions from Beijing using differential column measurements},
+  title      = {Lagrangian inversion of anthropogenic $CO_2$ emissions from Beijing using differential column measurements},
   url        = {https://dx.doi.org/10.1088/1748-9326/ac7477},
   volume     = {17},
   year       = {2022},
@@ -124,7 +124,7 @@
   journal    = {Atmospheric Measurement Techniques},
   number     = {9},
   pages      = {4751--4771},
-  title      = {Intercomparison of atmospheric CO2 and CH4 abundances on regional scales in boreal areas using Copernicus Atmosphere Monitoring Service (CAMS) analysis, COllaborative Carbon Column Observing Network (COCCON) spectrometers, and Sentinel-5 Precursor satellite observations},
+  title      = {Intercomparison of atmospheric $CO_2$ and $CH_4$ abundances on regional scales in boreal areas using Copernicus Atmosphere Monitoring Service (CAMS) analysis, COllaborative Carbon Column Observing Network (COCCON) spectrometers, and Sentinel-5 Precursor satellite observations},
   url        = {https://amt.copernicus.org/articles/13/4751/2020/},
   volume     = {13},
   year       = {2020},
@@ -139,7 +139,7 @@
   journal    = {Atmospheric Chemistry and Physics},
   number     = {14},
   pages      = {10939--10963},
-  title      = {The CO$_{2}$ integral emission by the megacity of St Petersburg as quantified from ground-based FTIR measurements combined with dispersion modelling},
+  title      = {The $CO_2$ integral emission by the megacity of St Petersburg as quantified from ground-based FTIR measurements combined with dispersion modelling},
   url        = {https://acp.copernicus.org/articles/21/10939/2021/},
   volume     = {21},
   year       = {2021},
@@ -197,7 +197,7 @@
   doi        = {10.3389/frsen.2021.775805},
   issn       = {2673-6187},
   journal    = {Frontiers in Remote Sensing},
-  title      = {Versatile and Targeted Validation of Space-Borne XCO2, XCH4 and XCO Observations by Mobile Ground-Based Direct-Sun Spectrometers},
+  title      = {Versatile and Targeted Validation of Space-Borne $XCO_2$, $XCH_4$ and $XCO$ Observations by Mobile Ground-Based Direct-Sun Spectrometers},
   url        = {https://www.frontiersin.org/articles/10.3389/frsen.2021.775805},
   volume     = {2},
   year       = {2022},
@@ -227,7 +227,7 @@
   journal    = {Atmospheric Measurement Techniques},
   number     = {12},
   pages      = {5023--5038},
-  title      = {Accurate mobile remote sensing of XCO$_{2}$ and XCH$_{4}$ latitudinal transects from aboard a research vessel},
[…]
