### 2023-04-04 tum-esm/pyra#158 @patrickjaigner
Dear @nmstreethran,
thank you for the detailed feedback.

We updated all your points and fully rewrote the bibliography by downloading the Bibtex files from the journals directly.

For the definition of UI and CLI in the summary section, we were unsure how to improve it. Currently, they are introduced like this: Pyra Command Line Interface (Pyra CLI), and Pyra User Interface (Pyra UI). Maybe you have an idea on how to improve this?
