### 2023-04-06 tum-esm/pyra#158 @nmstreethran
Hi @patrickjaigner, thank you for addressing my points.

> For the definition of UI and CLI in the summary section, we were unsure how to improve it. Currently, they are introduced like this: Pyra Command Line Interface (Pyra CLI), and Pyra User Interface (Pyra UI). Maybe you have an idea on how to improve this?

I'll review your paper again and get back to you on this. I have some comments on the installation below.

I tried to set-up Pyra as detailed in the user guide on my system running Windows 11 and seem to have run into some issues.

`python run.py` executes as expected, but `status` shows no local Pyra versions and the CLI version points to `undefined`. The `Documents\pyra` directory just has an empty folder `ui-installers`. Only `colorama`, `pip`, and `setuptools` are installed in the system interpreter.

```
PS C:\Users\geog\Documents\pyra-setup-tool> python run.py
------------------------------------------------------------------------------------------------------------------------
Welcome to the pyra-setup-tool! Instruction formatting:
  * checkpoints = green
  * expecting user input = yellow
------------------------------------------------------------------------------------------------------------------------
Please DO NOT use a virtual environment for PYRA! Use the system interpreter available via the command 'python'. Currently using the interpreter 'C:\Users\geog\AppData\Local\Programs\Python\Python310\python.exe'
Run 'which python' in another shell. Are system-interpreter and the current one identical? (Y | n) Y
Python version 3.10.10 is supported
Running command "poetry --version"" in directory C:\Users\geog\Documents\pyra-setup-tool"
Found poetry!
Running command "tar --version"" in directory C:\Users\geog\Documents\pyra-setup-tool"
Found tar!
Running command "git --version"" in directory C:\Users\geog\Documents\pyra-setup-tool"
Found git!
Running command "gh --version"" in directory C:\Users\geog\Documents\pyra-setup-tool"
Found github cli!
Running command "git pull"" in directory C:\Users\geog\Documents\pyra-setup-tool"
Setup tool is up to date.
Created directory 'C:\Users\geog\Documents\pyra'
Created directory 'C:\Users\geog\Documents\pyra\ui-installers'
Successfully initialized local environment inside "C:\Users\geog\Documents\pyra"
------------------------------------------------------------------------------------------------------------------------
Enter a command (help | status | install | rollback | remove | exit) status
Did not find any local pyra versions.
Running command "gh release list --repo tum-esm/pyra"" in directory C:\Users\geog\Documents\pyra-setup-tool"
Remote pyra versions: 4.0.7, 4.0.6, 4.0.5
Pyra CLI found in environment PATH variable: Running command "echo %PATH%"" in directory C:\Users\geog\Documents\pyra-setup-tool"
No
Pyra CLI is pointing to version: undefined
------------------------------------------------------------------------------------------------------------------------
Enter a command (help | status | install | rollback | remove | exit)
```

So, I tried to run `install` but GitHub CLI has a bug (https://github.com/cli/cli/issues/7283) that's causing the installation to fail.

```
Enter a command (help | status | install | rollback | remove | exit) install
Running command "gh release list --repo tum-esm/pyra"" in directory C:\Users\geog\Documents\pyra-setup-tool"
Which version should be installed? (4.0.7 | 4.0.6 | 4.0.5) 4.0.7
Running command "gh release download --repo tum-esm/pyra --archive=tar.gz v4.0.7"" in directory C:\Users\geog\Documents\pyra"
Exception occured!
Traceback (most recent call last):
  File "C:\Users\geog\Documents\pyra-setup-tool\run.py", line 5, in <module>
    main.run()
  File "C:\Users\geog\Documents\pyra-setup-tool\src\main.py", line 107, in run
    raise e
  File "C:\Users\geog\Documents\pyra-setup-tool\src\main.py", line 88, in run
    commands.install.run()
  File "C:\Users\geog\Documents\pyra-setup-tool\src\commands\install.py", line 24, in run
    tasks.manage_local_files.download_version(version_to_be_installed)
  File "C:\Users\geog\Documents\pyra-setup-tool\src\tasks\manage_local_files.py", line 16, in download_version
    shell_utils.run_shell_command(
  File "C:\Users\geog\Documents\pyra-setup-tool\src\utils\shell_utils.py", line 33, in run_shell_command
    assert p.returncode == 0, (
AssertionError: command 'gh release download --repo tum-esm/pyra --archive=tar.gz v4.0.7' failed with exit code 2: stderr = 'panic: runtime error: invalid memory address or nil pointer dereference
[signal 0xc0000005 code=0x0 addr=0x38 pc=0xc59337]

goroutine 9 [running]:
github.com/zalando/go-keyring.windowsKeychain.Get({}, {0xc00034c010?, 0x3?}, {0x0?, 0xa?})
        /home/runner/go/pkg/mod/github.com/zalando/go-keyring@v0.2.2/keyring_windows.go:21 +0x97
github.com/zalando/go-keyring.Get(...)
        /home/runner/go/pkg/mod/github.com/zalando/go-keyring@v0.2.2/keyring.go:37
github.com/cli/cli/v2/internal/config.(*AuthConfig).TokenFromKeyring(0x178217e?, {0x178217e?, 0x0?})
        /home/runner/work/cli/cli/internal/config/config.go:171 +0x53
github.com/cli/cli/v2/internal/config.(*AuthConfig).Token(0xc00006eea0, {0x178217e, 0xa})
        /home/runner/work/cli/cli/internal/config/config.go:135 +0x68
github.com/cli/cli/v2/api.AddAuthTokenHeader.func1(0xc0007de100)
        /home/runner/work/cli/cli/api/http_client.go:96 +0xa9
github.com/cli/cli/v2/api.funcTripper.RoundTrip(...)
        /home/runner/work/cli/cli/api/http_client.go:125
github.com/cli/cli/v2/api.AddASCIISanitizer.func1(0x2?)
        /home/runner/work/cli/cli/api/sanitize_ascii.go:26 +0x3a
github.com/cli/cli/v2/api.funcTripper.RoundTrip(...)
        /home/runner/work/cli/cli/api/http_client.go:125
github.com/cli/cli/v2/api.ExtractHeader.func1.1(0x0?)
        /home/runner/work/cli/cli/api/http_client.go:109 +0x47
github.com/cli/cli/v2/api.funcTripper.RoundTrip(...)
        /home/runner/work/cli/cli/api/http_client.go:125
net/http.send(0xc0007de100, {0x26b4380, 0xc0004b6050}, {0x1730c80?, 0x2287334ef01?, 0x0?})
        /opt/hostedtoolcache/go/1.19.7/x64/src/net/http/client.go:251 +0x5f7
net/http.(*Client).send(0xc0008521e0, 0xc0007de100, {0x0?, 0xc000883c50?, 0x0?})
        /opt/hostedtoolcache/go/1.19.7/x64/src/net/http/client.go:175 +0x9b
net/http.(*Client).do(0xc0008521e0, 0xc0007de100)
        /opt/hostedtoolcache/go/1.19.7/x64/src/net/http/client.go:715 +0x8fc
net/http.(*Client).Do(...)
        /opt/hostedtoolcache/go/1.19.7/x64/src/net/http/client.go:581
github.com/cli/cli/v2/pkg/cmd/release/shared.fetchReleasePath({0x26bc4d8, 0xc0003e0040}, 0xc000883fa0?, {0xc00049f6c0?, 0x3?}, {0xc0004ce000, 0x27})
        /home/runner/work/cli/cli/pkg/cmd/release/shared/fetch.go:206 +0xe9
github.com/cli/cli/v2/pkg/cmd/release/shared.FetchRelease.func1()
        /home/runner/work/cli/cli/pkg/cmd/release/shared/fetch.go:140 +0x1a7
created by github.com/cli/cli/v2/pkg/cmd/release/shared.FetchRelease
        /home/runner/work/cli/cli/pkg/cmd/release/shared/fetch.go:138 +0x15e
'
PS C:\Users\geog\Documents\pyra-setup-tool>
```

I downgraded GitHub CLI from `2.26.1` to `2.25.1` but the issue still persists. Which version did you use?

I'm not sure why Pyra does not install when `python run.py` is executed. Am I missing any steps, or could this be due to the CLI bug?

Software versions:

```
Python 3.10.10
Poetry (version 1.4.2)
git version 2.40.0.windows.1
gh version 2.26.1 (2023-04-04)
```
