### 2023-04-11 tum-esm/pyra#158 @nmstreethran
@patrickjaigner I'm afraid I'm still having issues. See below. There still isn't a local Pyra version after running the setup tool.

Any idea why this isn't working on my side? I'm guessing the curl command piped into tar isn't working as expected.

```
PS C:\Users\geog> cd .\Documents\
PS C:\Users\geog\Documents> git clone https://github.com/tum-esm/pyra-setup-tool
Cloning into 'pyra-setup-tool'...
remote: Enumerating objects: 631, done.
remote: Total 631 (delta 12), reused 10 (delta 10), pack-reused 600
Receiving objects: 100% (631/631), 114.16 KiB | 578.00 KiB/s, done.
Resolving deltas: 100% (363/363), done.
PS C:\Users\geog\Documents> cd .\pyra-setup-tool\
PS C:\Users\geog\Documents\pyra-setup-tool> python run.py
------------------------------------------------------------------------------------------------------------------------
Welcome to the pyra-setup-tool! Instruction formatting:
  * checkpoints = green
  * expecting user input = yellow
------------------------------------------------------------------------------------------------------------------------
Please DO NOT use a virtual environment for PYRA! Use the system interpreter available via the command 'python'. Currently using the interpreter 'C:\Users\geog\AppData\Local\Programs\Python\Python310\python.exe'
Run 'which python' in another shell. Are system-interpreter and the current one identical? (Y | n) Y
Python version 3.10.10 is supported
Running command "poetry --version" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Found poetry!
Running command "curl --version" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Found curl!
Running command "tar --version" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Found tar!
Running command "git --version" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Found git!
Running command "git pull" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Setup tool is up to date.
Created directory 'C:\Users\geog\Documents\pyra'
Created directory 'C:\Users\geog\Documents\pyra\ui-installers'
Successfully initialized local environment inside "C:\Users\geog\Documents\pyra"
------------------------------------------------------------------------------------------------------------------------
Enter a command (help | status | install | rollback | remove | exit) status
Did not find any local pyra versions.
Running command "curl --request GET --url "https://api.github.com/repos/tum-esm/pyra/releases" --header "Accept: application/vnd.github+json" --header "X-GitHub-Api-Version: 2022-11-28"" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Remote pyra versions: 4.0.7, 4.0.6
Running command "echo %PATH%" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Pyra CLI found in environment PATH variable: No
Pyra CLI is pointing to version: undefined
------------------------------------------------------------------------------------------------------------------------
Enter a command (help | status | install | rollback | remove | exit) install
Running command "curl --request GET --url "https://api.github.com/repos/tum-esm/pyra/releases" --header "Accept: application/vnd.github+json" --header "X-GitHub-Api-Version: 2022-11-28"" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Which version should be installed? (4.0.7 | 4.0.6) 4.0.7
Running command "curl -L https://github.com/tum-esm/pyra/archive/refs/tags/v4.0.7.tar.gz | tar zx" in directory "C:\Users\geog\Documents\pyra"
Exception occured!
Traceback (most recent call last):
  File "C:\Users\geog\Documents\pyra-setup-tool\run.py", line 5, in <module>
    main.run()
  File "C:\Users\geog\Documents\pyra-setup-tool\src\main.py", line 115, in run
    raise e
  File "C:\Users\geog\Documents\pyra-setup-tool\src\main.py", line 96, in run
    commands.install.run()
  File "C:\Users\geog\Documents\pyra-setup-tool\src\commands\install.py", line 41, in run
    tasks.download_version.download_version(version_to_be_installed)
  File "C:\Users\geog\Documents\pyra-setup-tool\src\tasks\download_version.py", line 12, in download_version
    utils.shell_utils.run_shell_command(
  File "C:\Users\geog\Documents\pyra-setup-tool\src\utils\shell_utils.py", line 41, in run_shell_command
    p.returncode == 0
AssertionError: command 'curl -L https://github.com/tum-esm/pyra/archive/refs/tags/v4.0.7.tar.gz | tar zx' failed with exit code 1: stderr = '  % Total    % Received % Xferd  Average Speed   Time    Time     Time  Current
                                 Dload  Upload   Total   Spent    Left  Speed
  0     0    0     0    0     0      0      0 --:--:-- --:--:-- --:--:--     0tar: Error opening archive: Failed to open '\\.\tape0'
  0 5215k    0 12842    0     0  17279      0  0:05:09 --:--:--  0:05:09 17279
curl: (23) Failure writing output to destination
'
PS C:\Users\geog\Documents\pyra-setup-tool>
```
