### 2023-04-11 tum-esm/pyra#158 @nmstreethran
@patrickjaigner I just tried it with Git Bash and it works now. I think it would be good if you state which shell to use in the docs (and also that the software is meant to be run on Windows). I mistakenly used PowerShell after seeing the tip about running `.ps1` scripts in the setup guide.

```
geog@DELL-LATITUDE-5410 MINGW64 ~
$ cd Documents/pyra-setup-tool/

geog@DELL-LATITUDE-5410 MINGW64 ~/Documents/pyra-setup-tool (main)
$ python run.py
------------------------------------------------------------------------------------------------------------------------
Welcome to the pyra-setup-tool! Instruction formatting:
  * checkpoints = green
  * expecting user input = yellow
------------------------------------------------------------------------------------------------------------------------
Please DO NOT use a virtual environment for PYRA! Use the system interpreter available via the command 'python'. Currently using the interpreter 'C:\Users\geog\AppData\Local\Programs\Python\Python310\python.exe'
Run 'which python' in another shell. Are system-interpreter and the current one identical? (Y | n) Y
Python version 3.10.10 is supported
Running command "poetry --version" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Found poetry!
Running command "curl --version" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Found curl!
Running command "tar --version" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Found tar!
Running command "git --version" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Found git!
Running command "git pull" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Setup tool is up to date.
Created directory 'C:\Users\geog\Documents\pyra'
Created directory 'C:\Users\geog\Documents\pyra\ui-installers'
Successfully initialized local environment inside "C:\Users\geog\Documents\pyra"
------------------------------------------------------------------------------------------------------------------------
Enter a command (help | status | install | rollback | remove | exit) status
Did not find any local pyra versions.
Running command "curl --request GET --url "https://api.github.com/repos/tum-esm/pyra/releases" --header "Accept: application/vnd.github+json" --header "X-GitHub-Api-Version: 2022-11-28"" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Remote pyra versions: 4.0.7, 4.0.6
Running command "echo %PATH%" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Pyra CLI found in environment PATH variable: No
Pyra CLI is pointing to version: undefined
------------------------------------------------------------------------------------------------------------------------
Enter a command (help | status | install | rollback | remove | exit) install
Running command "curl --request GET --url "https://api.github.com/repos/tum-esm/pyra/releases" --header "Accept: application/vnd.github+json" --header "X-GitHub-Api-Version: 2022-11-28"" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Which version should be installed? (4.0.7 | 4.0.6) 4.0.7
Running command "curl -L https://github.com/tum-esm/pyra/archive/refs/tags/v4.0.7.tar.gz | tar zx" in directory "C:\Users\geog\Documents\pyra"
Downloaded code
Running command "curl -L -O "https://github.com/tum-esm/pyra/releases/download/v4.0.7/Pyra.UI_4.0.7_x64_en-US.msi"" in directory "C:\Users\geog\Documents\pyra\ui-installers"
Downloaded UI
Running command "poetry config virtualenvs.create false" in directory "C:\Users\geog\Documents\pyra\pyra-4.0.7"
Running command "poetry env use system" in directory "C:\Users\geog\Documents\pyra\pyra-4.0.7"
Running command "poetry install" in directory "C:\Users\geog\Documents\pyra\pyra-4.0.7"
Skipping virtualenv creation, as specified in config file.
Installing dependencies from lock file

Package operations: 45 installs, 1 update, 0 removals

  • Installing pycparser (2.21)
  • Installing cffi (1.15.1)
  • Installing attrs (22.2.0)
  • Installing bcrypt (4.0.1)
  • Installing cryptography (39.0.1)
  • Installing exceptiongroup (1.1.0)
  • Installing iniconfig (2.0.0)
  • Installing numpy (1.24.2)
  • Installing packaging (23.0)
  • Installing pluggy (1.0.0)
  • Installing pynacl (1.5.0)
  • Installing six (1.16.0)
  • Installing tomli (2.0.1)
  • Installing click (8.1.3)
  • Installing coverage (7.1.0)
  • Installing mypy-extensions (1.0.0)
  • Installing invoke (1.7.3)
  • Installing ordered-set (4.1.0)
  • Installing pathlib2 (2.3.7.post1)
  • Installing platformdirs (3.0.0)
  • Installing pytest (7.2.1)
  • Installing orjson (3.8.6)
  • Installing pathspec (0.11.0)
  • Installing paramiko (3.0.0)
  • Installing pyerfa (2.0.0.1)
  • Installing pyyaml (6.0)
  • Updating setuptools (65.5.0 -> 67.2.0)
  • Installing typing-extensions (4.4.0)
  • Installing astropy (5.2.1)
  • Installing cerberus (1.3.4)
  • Installing black (22.12.0)
  • Installing filelock (3.9.0)
  • Installing deepdiff (6.2.3)
  • Installing opencv-python (4.7.0.68)
  • Installing mypy (0.991)
  • Installing fabric (2.7.1)
  • Installing psutil (5.9.1)
  • Installing pydantic (1.10.4)
  • Installing jdcal (1.4.1)
  • Installing pytest-cov (3.0.0)
  • Installing python-snap7 (1.3)
  • Installing pywin32 (303)
  • Installing tqdm (4.64.1)
  • Installing types-invoke (1.7.3.17)
  • Installing types-paramiko (2.12.0.3)
  • Installing types-psutil (5.9.5.6)
Connection pool is full, discarding connection: pypi.org. Connection pool size: 10
Installed code dependencies
Please install the UI using the installer that opens now
Running command "msiexec /i C:\Users\geog\Documents\pyra\ui-installers\Pyra.UI_4.0.7_x64_en-US.msi /qf" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Installed the UI
Updated the link in pyra-cli.bat
Running command "echo %PATH%" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Make the "pyra-cli" command available, by adding "C:\Users\geog\Documents\pyra" to your "user environment variables". See the pyra setup docs. (ok) ok
Created desktop shortcut to code directory
Skipping migration, no available versions to migrate from
done!
------------------------------------------------------------------------------------------------------------------------
Enter a command (help | status | install | rollback | remove | exit) status
Local pyra versions: 4.0.7
Running command "curl --request GET --url "https://api.github.com/repos/tum-esm/pyra/releases" --header "Accept: application/vnd.github+json" --header "X-GitHub-Api-Version: 2022-11-28"" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Remote pyra versions: 4.0.7, 4.0.6
Running command "echo %PATH%" in directory "C:\Users\geog\Documents\pyra-setup-tool"
Pyra CLI found in environment PATH variable: No
Pyra CLI is pointing to version: 4.0.7
------------------------------------------------------------------------------------------------------------------------
Enter a command (help | status | install | rollback | remove | exit)
```
