### 2023-04-12 tum-esm/pyra#158 @nmstreethran
@patrickjaigner

I checked Pyra UI and Pyra CLI and everything is working well. Based on this and your suggestions in https://github.com/openjournals/joss-reviews/issues/5131#issuecomment-1493958647 re: the functionality, I've checked the remaining items in my checklist.

Some minor comments on your paper:

- There are still some issues with capitalisation in the bibliography (see the [PDF](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05131/joss.05131/10.21105.joss.05131.pdf)). I've attached a file ([updated.txt](https://github.com/tum-esm/pyra/files/11210936/updated.txt)) with the suggested changes which you can review and incorporate into your BibTeX file.
- Under Software Dependencies, can you update the bit about the Pyra Setup Tool (since you're no longer using GitHub CLI)?
- Under Software Dependencies, you state that the UI is a Windows 10 application. I'm not sure if the Windows version is relevant here... I'm guessing I should be able to use it on Win 8/10/11 without any issues.
- In the Summary, under Pyra CLI, you state "The core commands allow direct interaction with the Pyra Core process and manage that only one instance of the Pyra Core process is running at the same time". If I understand this correctly, I think it will be clearer if you state "... **ensure** that only one instance of the Pyra Core process is running at **a** time."

Additionally:

- The link to the contributor guide in the repo's README is broken.
- Another thing that you may want to consider is adding a link to the docs and/or GitHub repo to the Pyra UI and CLI so users can easily get support. Note that this is just a suggestion and is not needed for the JOSS review.
- In the setup guide, I suggest changing the colorama installation command to `python -m pip install colorama` (which is [recommended](https://docs.python.org/3/installing/index.html)).
