+++ README.md
@@ -1,8 +1,8 @@
 # PYRA
 
-**Documentation on: [pyra.esm.ei.tum.de/docs](https://pyra.esm.ei.tum.de/docs)**
+**Documentation: [pyra.esm.ei.tum.de/docs](https://pyra.esm.ei.tum.de/docs)**
 
-**Contributor Guide on: [https://pyra.esm.ei.tum.de/docs/category/contributor-guide](https://pyra.esm.ei.tum.de/docs/category/contributor-guide)**
+**Contributor Guide: [https://pyra.esm.ei.tum.de/docs/contributor-guide/becoming-a-contributor](https://pyra.esm.ei.tum.de/docs/contributor-guide/becoming-a-contributor)**
 
 ![GitHub Workflow Status](https://img.shields.io/github/actions/workflow/status/tum-esm/pyra/test-python-codebase-on-main.yml?branch=main&color=86efac&label=python%20tests%20on%20main%20branch&style=for-the-badge)
 ![GitHub Workflow Status](https://img.shields.io/github/actions/workflow/status/tum-esm/pyra/test-typescript-codebase-on-main.yml?branch=main&color=86efac&label=typescript%20tests%20on%20main%20branch&style=for-the-badge)
