### 2023-04-12 b4f0f66
Update bibliography and paper #158
