+++ joss/paper.bib
@@ -4,7 +4,7 @@
   journal    = {Atmospheric Chemistry and Physics},
   number     = {17},
   pages      = {13131--13147},
-  title      = {Assessing urban methane emissions using column-observing portable Fourier transform infrared (FTIR) spectrometers and a novel Bayesian inversion framework},
+  title      = {Assessing urban methane emissions using column-observing portable {Fourier} transform infrared (FTIR) spectrometers and a novel {Bayesian} inversion framework},
   url        = {https://acp.copernicus.org/articles/21/13131/2021/},
   volume     = {21},
   year       = {2021},
@@ -18,7 +18,7 @@
   journal    = {Atmospheric Measurement Techniques},
   number     = {2},
   pages      = {1111--1126},
-  title      = {MUCCnet: Munich Urban Carbon Column network},
+  title      = {{MUCCnet: Munich Urban Carbon Column network}},
   url        = {https://amt.copernicus.org/articles/14/1111/2021/},
   volume     = {14},
   year       = {2021},
@@ -33,7 +33,7 @@
   journal    = {Atmospheric Chemistry and Physics},
   number     = {6},
   pages      = {3683--3696},
-  title      = {Methane emissions from the Munich Oktoberfest},
+  title      = {Methane emissions from the {Munich Oktoberfest}},
   url        = {https://acp.copernicus.org/articles/20/3683/2020/},
   volume     = {20},
   year       = {2020},
@@ -48,7 +48,7 @@
   journal    = {Atmospheric Chemistry and Physics},
   number     = {9},
   pages      = {5859--5876},
-  title      = {Observational constraints on methane emissions from Polish coal mines using a ground-based remote sensing network},
+  title      = {Observational constraints on methane emissions from {Polish} coal mines using a ground-based remote sensing network},
   url        = {https://acp.copernicus.org/articles/22/5859/2022/},
   volume     = {22},
   year       = {2022},
@@ -63,7 +63,7 @@
   journal    = {Atmospheric Chemistry and Physics},
   number     = {1},
   pages      = {295--317},
-  title      = {Quantification of $CH_4$ emissions from waste disposal sites near the city of Madrid using ground- and space-based observations of COCCON, IASI},
+  title      = {Quantification of $CH_4$ emissions from waste disposal sites near the city of {Madrid} using ground- and space-based observations of COCCON, IASI},
   url        = {https://acp.copernicus.org/articles/22/295/2022/},
   volume     = {22},
   year       = {2022},
@@ -78,7 +78,7 @@
   journal    = {Atmospheric Measurement Techniques},
   number     = {7},
   pages      = {3059--3068},
-  title      = {Application of portable FTIR spectrometers for detecting greenhouse gas emissions of the major city Berlin},
+  title      = {Application of portable FTIR spectrometers for detecting greenhouse gas emissions of the major city {Berlin}},
   url        = {https://amt.copernicus.org/articles/8/3059/2015/},
   volume     = {8},
   year       = {2015},
@@ -95,7 +95,7 @@
   number     = {7},
   pages      = {075001},
   publisher  = {IOP Publishing},
-  title      = {Lagrangian inversion of anthropogenic $CO_2$ emissions from Beijing using differential column measurements},
+  title      = {Lagrangian inversion of anthropogenic $CO_2$ emissions from {Beijing} using differential column measurements},
   url        = {https://dx.doi.org/10.1088/1748-9326/ac7477},
[…]
