### 2023-04-27 #5131 @kthyng
Hi @patrickjaigner! Not sure why the reference checker won't work. But can you update the metadata in your Zenodo archive so the title and author list exactly match your JOSS paper?
