+++ README.md
@@ -1,4 +1,4 @@
-# PYRA: Automated EM27/SUN Greenhouse Gas Measurement Software
+# Pyra: Automated EM27/SUN Greenhouse Gas Measurement Software
 
 **Documentation: [pyra.esm.ei.tum.de/docs](https://pyra.esm.ei.tum.de/docs)**
 
